Ouroboros.Consensus.Byron.Ledger.Forge

Documentation

forgeByronBlock

Arguments

:: HasCallStack

=> TopLevelConfig ByronBlock

-> TxOverrides ByronBlock
How to override max tx capacity defined by ledger

-> BlockNo
Current block number

-> SlotNo
Current slot number

-> TickedLedgerState ByronBlock
Current ledger

-> [Validated (GenTx ByronBlock)]
Txs to consider adding in the block

-> PBftIsLeader PBftByronCrypto
Leader proof (IsLeader)

-> ByronBlock

forgeRegularBlock

Arguments

:: HasCallStack

=> BlockConfig ByronBlock

-> TxOverrides ByronBlock
How to override max tx capacity defined by ledger

-> BlockNo
Current block number

-> SlotNo
Current slot number

-> TickedLedgerState ByronBlock
Current ledger

-> [Validated (GenTx ByronBlock)]
Txs to consider adding in the block

-> PBftIsLeader PBftByronCrypto
Leader proof (IsLeader)

-> ByronBlock

For testing purposes

forgeEBB

Arguments

:: BlockConfig ByronBlock

-> SlotNo
Current slot

-> BlockNo
Current block number

-> ChainHash ByronBlock
Previous hash

-> ByronBlock