Ouroboros.Consensus.Byron.Ledger.Mempool

Description

Byron mempool integration

Mempool integration

data family GenTx blk

data family TxId tx

data family Validated x

Transaction IDs

byronIdDlg :: ACertificate ByteString -> CertificateId

byronIdProp :: AProposal ByteString -> UpId

byronIdTx :: ATxAux ByteString -> TxId

byronIdVote :: AVote ByteString -> VoteId

Serialisation

decodeByronApplyTxError :: Decoder s (ApplyTxErr ByronBlock)

decodeByronGenTx :: Decoder s (GenTx ByronBlock)

The ByteString annotation will be the canonical encoding.

While the new implementation does not care about canonical encodings, the old one does. When a generalised transaction arrives that is not in its canonical encoding (only the ATxAux of the ByronTx can be produced by nodes that are not under our control), the old implementation will reject it. Therefore, we need to reject them too. See #905.

We use the ledger to check for canonical encodings: the ledger will check whether the signed hash of the transaction (in the case of a ATxAux, the transaction witness) matches the annotated bytestring. Is therefore important that the annotated bytestring be the canonical encoding, not the original, possibly non-canonical encoding.

decodeByronGenTxId :: Decoder s (GenTxId ByronBlock)

encodeByronApplyTxError :: ApplyTxErr ByronBlock -> Encoding

encodeByronGenTx :: GenTx ByronBlock -> Encoding

encodeByronGenTxId :: GenTxId ByronBlock -> Encoding

Low-level API (primarily for testing)

fromMempoolPayload :: AMempoolPayload ByteString -> GenTx ByronBlock

toMempoolPayload :: GenTx ByronBlock -> AMempoolPayload ByteString

Auxiliary functions

countByronGenTxs :: ByronBlock -> Word64

Count all (generalized) transactions in the block

Orphan instances

HasTxs ByronBlock

Instance details

Methods

extractTxs :: ByronBlock -> [GenTx ByronBlock]

LedgerSupportsMempool ByronBlock

Instance details

Methods

txInvariant :: GenTx ByronBlock -> Bool

applyTx :: LedgerConfig ByronBlock -> WhetherToIntervene -> SlotNo -> GenTx ByronBlock -> TickedLedgerState ByronBlock -> Except (ApplyTxErr ByronBlock) (TickedLedgerState ByronBlock, Validated (GenTx ByronBlock))

reapplyTx :: LedgerConfig ByronBlock -> SlotNo -> Validated (GenTx ByronBlock) -> TickedLedgerState ByronBlock -> Except (ApplyTxErr ByronBlock) (TickedLedgerState ByronBlock)

txsMaxBytes :: TickedLedgerState ByronBlock -> Word32

txInBlockSize :: GenTx ByronBlock -> Word32

txForgetValidated :: Validated (GenTx ByronBlock) -> GenTx ByronBlock

TxLimits ByronBlock

Instance details

Associated Types

type TxMeasure ByronBlock

Methods

txMeasure :: Validated (GenTx ByronBlock) -> TxMeasure ByronBlock

txsBlockCapacity :: Ticked (LedgerState ByronBlock) -> TxMeasure ByronBlock

ShowProxy ApplyMempoolPayloadErr

Instance details

Methods

showProxy :: Proxy ApplyMempoolPayloadErr -> String

ShowProxy (GenTx ByronBlock)

Instance details

Methods

showProxy :: Proxy (GenTx ByronBlock) -> String

ShowProxy (TxId (GenTx ByronBlock))

Instance details

Methods

showProxy :: Proxy (TxId (GenTx ByronBlock)) -> String

Generic (Validated (GenTx ByronBlock))

Instance details

Associated Types

type Rep (Validated (GenTx ByronBlock)) :: Type -> Type

Methods

from :: Validated (GenTx ByronBlock) -> Rep (Validated (GenTx ByronBlock)) x

to :: Rep (Validated (GenTx ByronBlock)) x -> Validated (GenTx ByronBlock)

Generic (GenTx ByronBlock)

Instance details

Associated Types

type Rep (GenTx ByronBlock) :: Type -> Type

Methods

from :: GenTx ByronBlock -> Rep (GenTx ByronBlock) x

to :: Rep (GenTx ByronBlock) x -> GenTx ByronBlock

Show (Validated (GenTx ByronBlock))

Instance details

Methods

showsPrec :: Int -> Validated (GenTx ByronBlock) -> ShowS

show :: Validated (GenTx ByronBlock) -> String

showList :: [Validated (GenTx ByronBlock)] -> ShowS

Show (GenTx ByronBlock)

Instance details

Methods

showsPrec :: Int -> GenTx ByronBlock -> ShowS

show :: GenTx ByronBlock -> String

showList :: [GenTx ByronBlock] -> ShowS

Show (GenTxId ByronBlock)

Instance details

Methods

showsPrec :: Int -> GenTxId ByronBlock -> ShowS

show :: GenTxId ByronBlock -> String

showList :: [GenTxId ByronBlock] -> ShowS

Eq (Validated (GenTx ByronBlock))

Instance details

Methods

(==) :: Validated (GenTx ByronBlock) -> Validated (GenTx ByronBlock) -> Bool

(/=) :: Validated (GenTx ByronBlock) -> Validated (GenTx ByronBlock) -> Bool

Eq (GenTx ByronBlock)

Instance details

Methods

(==) :: GenTx ByronBlock -> GenTx ByronBlock -> Bool

(/=) :: GenTx ByronBlock -> GenTx ByronBlock -> Bool

Eq (TxId (GenTx ByronBlock))

Instance details

Methods

(==) :: TxId (GenTx ByronBlock) -> TxId (GenTx ByronBlock) -> Bool

(/=) :: TxId (GenTx ByronBlock) -> TxId (GenTx ByronBlock) -> Bool

Ord (TxId (GenTx ByronBlock))

Instance details

Methods

compare :: TxId (GenTx ByronBlock) -> TxId (GenTx ByronBlock) -> Ordering

(<) :: TxId (GenTx ByronBlock) -> TxId (GenTx ByronBlock) -> Bool

(<=) :: TxId (GenTx ByronBlock) -> TxId (GenTx ByronBlock) -> Bool

(>) :: TxId (GenTx ByronBlock) -> TxId (GenTx ByronBlock) -> Bool

(>=) :: TxId (GenTx ByronBlock) -> TxId (GenTx ByronBlock) -> Bool

max :: TxId (GenTx ByronBlock) -> TxId (GenTx ByronBlock) -> TxId (GenTx ByronBlock)

min :: TxId (GenTx ByronBlock) -> TxId (GenTx ByronBlock) -> TxId (GenTx ByronBlock)

NoThunks (Validated (GenTx ByronBlock))

Instance details

Methods

noThunks :: Context -> Validated (GenTx ByronBlock) -> IO (Maybe ThunkInfo)

wNoThunks :: Context -> Validated (GenTx ByronBlock) -> IO (Maybe ThunkInfo)

showTypeOf :: Proxy (Validated (GenTx ByronBlock)) -> String

NoThunks (GenTx ByronBlock)

Instance details

Methods

noThunks :: Context -> GenTx ByronBlock -> IO (Maybe ThunkInfo)

wNoThunks :: Context -> GenTx ByronBlock -> IO (Maybe ThunkInfo)

showTypeOf :: Proxy (GenTx ByronBlock) -> String

NoThunks (TxId (GenTx ByronBlock))

Instance details

Methods

noThunks :: Context -> TxId (GenTx ByronBlock) -> IO (Maybe ThunkInfo)

wNoThunks :: Context -> TxId (GenTx ByronBlock) -> IO (Maybe ThunkInfo)

showTypeOf :: Proxy (TxId (GenTx ByronBlock)) -> String

HasTxId (GenTx ByronBlock)

Instance details

Methods

txId :: GenTx ByronBlock -> TxId (GenTx ByronBlock)

Condense (GenTx ByronBlock)

Instance details

Methods

condense :: GenTx ByronBlock -> String

Condense (GenTxId ByronBlock)

Instance details

Methods

condense :: GenTxId ByronBlock -> String