Ouroboros.Consensus.Byron.Node

Documentation

newtype PBftSignatureThreshold

Constructors

PBftSignatureThreshold

Fields

getPBftSignatureThreshold :: Double

blockForgingByron :: Monad m => ProtocolParams ByronBlock -> [BlockForging m ByronBlock]

byronBlockForging :: Monad m => TxOverrides ByronBlock -> ByronLeaderCredentials -> BlockForging m ByronBlock

Secrets

data ByronLeaderCredentials

Credentials needed to produce blocks in the Byron era.

Constructors

ByronLeaderCredentials

Fields

blcSignKey :: SigningKey

blcDlgCert :: Certificate

blcCoreNodeId :: CoreNodeId

Only core nodes can produce blocks. The CoreNodeId is used to determine the order (round-robin) in which core nodes produce blocks.

blcLabel :: Text

Identifier for this set of credentials.

Useful when the node is running with multiple sets of credentials.

data ByronLeaderCredentialsError

mkByronLeaderCredentials :: Config -> SigningKey -> Certificate -> Text -> Either ByronLeaderCredentialsError ByronLeaderCredentials

Make the ByronLeaderCredentials, with a couple sanity checks:

That the block signing key and the delegation certificate match.

That the delegation certificate does correspond to one of the genesis keys from the genesis file.

mkPBftCanBeLeader :: ByronLeaderCredentials -> CanBeLeader (PBft PBftByronCrypto)

ProtocolInfo

data family ProtocolParams blk

defaultPBftSignatureThreshold :: PBftSignatureThreshold

See chapter 4.1 of https://hydra.iohk.io/job/Cardano/cardano-ledger-specs/byronChainSpec/latest/download-by-type/doc-pdf/blockchain-spec

mkByronConfig :: Config -> ProtocolVersion -> SoftwareVersion -> BlockConfig ByronBlock

protocolClientInfoByron :: EpochSlots -> ProtocolClientInfo ByronBlock

protocolInfoByron :: ProtocolParams ByronBlock -> ProtocolInfo ByronBlock

Orphan instances

BlockSupportsMetrics ByronBlock

Instance details

Methods

isSelfIssued :: BlockConfig ByronBlock -> Header ByronBlock -> WhetherSelfIssued

ConfigSupportsNode ByronBlock

Instance details

Methods

getSystemStart :: BlockConfig ByronBlock -> SystemStart

getNetworkMagic :: BlockConfig ByronBlock -> NetworkMagic

NodeInitStorage ByronBlock

Instance details

Methods

nodeImmutableDbChunkInfo :: StorageConfig ByronBlock -> ChunkInfo

nodeCheckIntegrity :: StorageConfig ByronBlock -> ByronBlock -> Bool

nodeInitChainDB :: IOLike m => StorageConfig ByronBlock -> InitChainDB m ByronBlock -> m ()

RunNode ByronBlock

Instance details