Ouroboros.Consensus.Byron.Node.Serialisation

Orphan instances

SerialiseNodeToClientConstraints ByronBlock

Instance details

SerialiseNodeToNodeConstraints ByronBlock

Instance details

Methods

estimateBlockSize :: Header ByronBlock -> SizeInBytes

SerialiseDiskConstraints ByronBlock

Instance details

HasBinaryBlockInfo ByronBlock

Instance details

Methods

getBinaryBlockInfo :: ByronBlock -> BinaryBlockInfo

SerialiseNodeToClient ByronBlock ApplyMempoolPayloadErr
ApplyTxErr ByronBlock

Instance details

Methods

encodeNodeToClient :: CodecConfig ByronBlock -> BlockNodeToClientVersion ByronBlock -> ApplyMempoolPayloadErr -> Encoding

decodeNodeToClient :: CodecConfig ByronBlock -> BlockNodeToClientVersion ByronBlock -> forall s. Decoder s ApplyMempoolPayloadErr

SerialiseNodeToClient ByronBlock SlotNo

Instance details

Methods

encodeNodeToClient :: CodecConfig ByronBlock -> BlockNodeToClientVersion ByronBlock -> SlotNo -> Encoding

decodeNodeToClient :: CodecConfig ByronBlock -> BlockNodeToClientVersion ByronBlock -> forall s. Decoder s SlotNo

SerialiseNodeToClient ByronBlock ByronBlock
CBOR-in-CBOR for the annotation. This also makes it compatible with the wrapped (Serialised) variant.

Instance details

Methods

encodeNodeToClient :: CodecConfig ByronBlock -> BlockNodeToClientVersion ByronBlock -> ByronBlock -> Encoding

decodeNodeToClient :: CodecConfig ByronBlock -> BlockNodeToClientVersion ByronBlock -> forall s. Decoder s ByronBlock

SerialiseNodeToNode ByronBlock ByronBlock
CBOR-in-CBOR for the annotation. This also makes it compatible with the wrapped (Serialised) variant.

Instance details

Methods

encodeNodeToNode :: CodecConfig ByronBlock -> BlockNodeToNodeVersion ByronBlock -> ByronBlock -> Encoding

decodeNodeToNode :: CodecConfig ByronBlock -> BlockNodeToNodeVersion ByronBlock -> forall s. Decoder s ByronBlock

EncodeDisk ByronBlock ByronBlock

Instance details

Methods

encodeDisk :: CodecConfig ByronBlock -> ByronBlock -> Encoding

ReconstructNestedCtxt Header ByronBlock

Instance details

Methods

reconstructPrefixLen :: proxy (Header ByronBlock) -> PrefixLen

reconstructNestedCtxt :: proxy (Header ByronBlock) -> ShortByteString -> SizeInBytes -> SomeSecond (NestedCtxt Header) ByronBlock

SerialiseNodeToClient ByronBlock (GenTx ByronBlock)
No CBOR-in-CBOR, because we check for canonical encodings, which means we can use the recomputed encoding for the annotation.

Instance details

Methods

encodeNodeToClient :: CodecConfig ByronBlock -> BlockNodeToClientVersion ByronBlock -> GenTx ByronBlock -> Encoding

decodeNodeToClient :: CodecConfig ByronBlock -> BlockNodeToClientVersion ByronBlock -> forall s. Decoder s (GenTx ByronBlock)

SerialiseNodeToClient ByronBlock (GenTxId ByronBlock)

Instance details

Methods

encodeNodeToClient :: CodecConfig ByronBlock -> BlockNodeToClientVersion ByronBlock -> GenTxId ByronBlock -> Encoding

decodeNodeToClient :: CodecConfig ByronBlock -> BlockNodeToClientVersion ByronBlock -> forall s. Decoder s (GenTxId ByronBlock)

SerialiseNodeToNode ByronBlock (Header ByronBlock)

Instance details

Methods

encodeNodeToNode :: CodecConfig ByronBlock -> BlockNodeToNodeVersion ByronBlock -> Header ByronBlock -> Encoding

decodeNodeToNode :: CodecConfig ByronBlock -> BlockNodeToNodeVersion ByronBlock -> forall s. Decoder s (Header ByronBlock)

SerialiseNodeToNode ByronBlock (GenTx ByronBlock)
No CBOR-in-CBOR, because we check for canonical encodings, which means we can use the recomputed encoding for the annotation.

Instance details

Methods

encodeNodeToNode :: CodecConfig ByronBlock -> BlockNodeToNodeVersion ByronBlock -> GenTx ByronBlock -> Encoding

decodeNodeToNode :: CodecConfig ByronBlock -> BlockNodeToNodeVersion ByronBlock -> forall s. Decoder s (GenTx ByronBlock)

SerialiseNodeToNode ByronBlock (GenTxId ByronBlock)

Instance details

Methods

encodeNodeToNode :: CodecConfig ByronBlock -> BlockNodeToNodeVersion ByronBlock -> GenTxId ByronBlock -> Encoding

decodeNodeToNode :: CodecConfig ByronBlock -> BlockNodeToNodeVersion ByronBlock -> forall s. Decoder s (GenTxId ByronBlock)

SerialiseNodeToNode ByronBlock (SerialisedHeader ByronBlock)

Instance details

Methods

encodeNodeToNode :: CodecConfig ByronBlock -> BlockNodeToNodeVersion ByronBlock -> SerialisedHeader ByronBlock -> Encoding

decodeNodeToNode :: CodecConfig ByronBlock -> BlockNodeToNodeVersion ByronBlock -> forall s. Decoder s (SerialisedHeader ByronBlock)

SerialiseResult ByronBlock (BlockQuery ByronBlock)

Instance details

Methods

encodeResult :: CodecConfig ByronBlock -> BlockNodeToClientVersion ByronBlock -> BlockQuery ByronBlock result -> result -> Encoding

decodeResult :: CodecConfig ByronBlock -> BlockNodeToClientVersion ByronBlock -> BlockQuery ByronBlock result -> forall s. Decoder s result

DecodeDisk ByronBlock (AnnTip ByronBlock)

Instance details

Methods

decodeDisk :: CodecConfig ByronBlock -> forall s. Decoder s (AnnTip ByronBlock)

DecodeDisk ByronBlock (LedgerState ByronBlock)

Instance details

Methods

decodeDisk :: CodecConfig ByronBlock -> forall s. Decoder s (LedgerState ByronBlock)

DecodeDisk ByronBlock (PBftState PBftByronCrypto)
ChainDepState (BlockProtocol ByronBlock)

Instance details

Methods

decodeDisk :: CodecConfig ByronBlock -> forall s. Decoder s (PBftState PBftByronCrypto)

EncodeDisk ByronBlock (AnnTip ByronBlock)

Instance details

Methods

encodeDisk :: CodecConfig ByronBlock -> AnnTip ByronBlock -> Encoding

EncodeDisk ByronBlock (LedgerState ByronBlock)

Instance details

Methods

encodeDisk :: CodecConfig ByronBlock -> LedgerState ByronBlock -> Encoding

EncodeDisk ByronBlock (PBftState PBftByronCrypto)
ChainDepState (BlockProtocol ByronBlock)

Instance details

Methods

encodeDisk :: CodecConfig ByronBlock -> PBftState PBftByronCrypto -> Encoding

SerialiseNodeToClient ByronBlock (SomeSecond BlockQuery ByronBlock)

Instance details

Methods

encodeNodeToClient :: CodecConfig ByronBlock -> BlockNodeToClientVersion ByronBlock -> SomeSecond BlockQuery ByronBlock -> Encoding

decodeNodeToClient :: CodecConfig ByronBlock -> BlockNodeToClientVersion ByronBlock -> forall s. Decoder s (SomeSecond BlockQuery ByronBlock)

SerialiseNodeToClient ByronBlock (Serialised ByronBlock)
Serialised uses CBOR-in-CBOR by default.

Instance details

Methods

encodeNodeToClient :: CodecConfig ByronBlock -> BlockNodeToClientVersion ByronBlock -> Serialised ByronBlock -> Encoding

decodeNodeToClient :: CodecConfig ByronBlock -> BlockNodeToClientVersion ByronBlock -> forall s. Decoder s (Serialised ByronBlock)

SerialiseNodeToNode ByronBlock (Serialised ByronBlock)
Serialised uses CBOR-in-CBOR by default.

Instance details

Methods

encodeNodeToNode :: CodecConfig ByronBlock -> BlockNodeToNodeVersion ByronBlock -> Serialised ByronBlock -> Encoding

decodeNodeToNode :: CodecConfig ByronBlock -> BlockNodeToNodeVersion ByronBlock -> forall s. Decoder s (Serialised ByronBlock)

DecodeDisk ByronBlock (ByteString -> ByronBlock)

Instance details

Methods

decodeDisk :: CodecConfig ByronBlock -> forall s. Decoder s (ByteString -> ByronBlock)

DecodeDiskDep (NestedCtxt Header) ByronBlock

Instance details

Methods

decodeDiskDep :: CodecConfig ByronBlock -> NestedCtxt Header ByronBlock a -> forall s. Decoder s (ByteString -> a)

DecodeDiskDepIx (NestedCtxt Header) ByronBlock

Instance details

Methods

decodeDiskDepIx :: CodecConfig ByronBlock -> Decoder s (SomeSecond (NestedCtxt Header) ByronBlock)

EncodeDiskDep (NestedCtxt Header) ByronBlock

Instance details

Methods

encodeDiskDep :: CodecConfig ByronBlock -> NestedCtxt Header ByronBlock a -> a -> Encoding

EncodeDiskDepIx (NestedCtxt Header) ByronBlock

Instance details

Methods

encodeDiskDepIx :: CodecConfig ByronBlock -> SomeSecond (NestedCtxt Header) ByronBlock -> Encoding