Ouroboros.Consensus.Byron.Protocol

Documentation

data PBftByronCrypto

genesisKeyCoreNodeId

Arguments

:: Config

-> VerKeyDSIGN ByronDSIGN
The genesis verification key

-> Maybe CoreNodeId

Determine the CoreNodeId for a code node, based on the genesis key it will sign blocks on behalf of.

In PBFT, the CoreNodeId index is determined by the 0-based position in the sort order of the genesis key hashes.

nodeIdToGenesisKey :: Config -> CoreNodeId -> Maybe KeyHash

Inverse of genesisKeyCoreNodeId