Ouroboros.Consensus.Cardano

The block type of the Cardano block chain

type CardanoBlock c = HardForkBlock (CardanoEras c)

The Cardano block.

Thanks to the pattern synonyms, you can treat this as a sum type with constructors BlockByron and BlockShelley.

f :: CardanoBlock c -> _
f (BlockByron   b) = _
f (BlockShelley s) = _
f (BlockAllegra a) = _
f (BlockMary    m) = _

Supported protocols

type ProtocolByron = HardForkProtocol '[ByronBlock]

type ProtocolShelley = HardForkProtocol '[ShelleyBlock StandardShelley]

type ProtocolCardano = HardForkProtocol '[ByronBlock, ShelleyBlock StandardShelley, ShelleyBlock StandardAllegra, ShelleyBlock StandardMary]

Abstract over the various protocols

data ProtocolParamsByron

Parameters needed to run Byron

Constructors

ProtocolParamsByron

Fields

byronGenesis ∷ Config

byronPbftSignatureThreshold ∷ Maybe PBftSignatureThreshold

byronProtocolVersion ∷ ProtocolVersion

byronSoftwareVersion ∷ SoftwareVersion

byronLeaderCredentials ∷ Maybe ByronLeaderCredentials

data ProtocolParamsShelley c (f ∷ Type → Type)

Parameters needed to run Shelley

Constructors

ProtocolParamsShelley

Fields

shelleyGenesis ∷ ShelleyGenesis (ShelleyEra c)

shelleyInitialNonce ∷ Nonce

The initial nonce, typically derived from the hash of Genesis config JSON file.

WARNING: chains using different values of this parameter will be mutually incompatible.

shelleyProtVer ∷ ProtVer

shelleyLeaderCredentials ∷ f (TPraosLeaderCredentials c)

data ProtocolParamsAllegra c (f ∷ Type → Type)

Parameters needed to run Allegra

Constructors

ProtocolParamsAllegra

Fields

allegraProtVer ∷ ProtVer

allegraLeaderCredentials ∷ f (TPraosLeaderCredentials c)

data ProtocolParamsMary c (f ∷ Type → Type)

Parameters needed to run Mary

Constructors

ProtocolParamsMary

Fields

maryProtVer ∷ ProtVer

maryLeaderCredentials ∷ f (TPraosLeaderCredentials c)

data ProtocolParamsTransition eraFrom eraTo

Parameters needed to transition between two eras.

The two eras are phantom type parameters of this type to avoid mixing up multiple ProtocolParamsTransitions

Constructors

ProtocolParamsTransition

Fields

transitionTrigger ∷ TriggerHardFork

data Protocol (m ∷ Type → Type) blk p where

Consensus protocol to use

Constructors

ProtocolByron ∷ ProtocolParamsByron → Protocol m ByronBlockHFC ProtocolByron
Run PBFT against the real Byron ledger

ProtocolShelley ∷ ProtocolParamsShelley StandardCrypto [] → Protocol m (ShelleyBlockHFC StandardShelley) ProtocolShelley
Run TPraos against the real Shelley ledger

ProtocolCardano ∷ ProtocolParamsByron → ProtocolParamsShelley StandardCrypto Maybe → ProtocolParamsAllegra StandardCrypto Maybe → ProtocolParamsMary StandardCrypto Maybe → ProtocolParamsTransition ByronBlock (ShelleyBlock StandardShelley) → ProtocolParamsTransition (ShelleyBlock StandardShelley) (ShelleyBlock StandardAllegra) → ProtocolParamsTransition (ShelleyBlock StandardAllegra) (ShelleyBlock StandardMary) → Protocol m (CardanoBlock StandardCrypto) ProtocolCardano
Run the protocols of the Cardano block

verifyProtocol ∷ Protocol m blk p → p :~: BlockProtocol blk

Data required to run a protocol

protocolInfo ∷ ∀ m blk p. IOLike m ⇒ Protocol m blk p → ProtocolInfo m blk

Data required to run the selected protocol

Evidence that we can run all the supported protocols

runProtocol ∷ Protocol m blk p → Dict (RunNode blk)

module Ouroboros.Consensus.Shelley.Node

module Ouroboros.Consensus.Byron.Node

module Ouroboros.Consensus.Protocol.PBFT

module Ouroboros.Consensus.Protocol.Abstract

Client support for nodes running a protocol

data ProtocolClient blk p where

Node client support for each consensus protocol.

This is like Protocol but for clients of the node, so with less onerous requirements than to run a node.

Constructors

ProtocolClientByron ∷ EpochSlots → ProtocolClient ByronBlockHFC ProtocolByron

ProtocolClientShelley ∷ ProtocolClient (ShelleyBlockHFC StandardShelley) ProtocolShelley

ProtocolClientCardano ∷ EpochSlots → ProtocolClient (CardanoBlock StandardCrypto) ProtocolCardano

protocolClientInfo ∷ ProtocolClient blk p → ProtocolClientInfo blk

Data required by clients of a node running the specified protocol.

runProtocolClient ∷ ProtocolClient blk p → Dict (RunNode blk)

Sanity check that we have the right class instances available

verifyProtocolClient ∷ ProtocolClient blk p → p :~: BlockProtocol blk

Sanity check that we have the right type combinations