Ouroboros.Consensus.Cardano.Node

Documentation

type CardanoHardForkConstraints c = (PraosCrypto c, PraosCrypto c, TranslateProto (TPraos c) (Praos c), ShelleyCompatible (TPraos c) (ShelleyEra c), LedgerSupportsProtocol (ShelleyBlock (TPraos c) (ShelleyEra c)), ShelleyCompatible (TPraos c) (AllegraEra c), LedgerSupportsProtocol (ShelleyBlock (TPraos c) (AllegraEra c)), ShelleyCompatible (TPraos c) (MaryEra c), LedgerSupportsProtocol (ShelleyBlock (TPraos c) (MaryEra c)), ShelleyCompatible (TPraos c) (AlonzoEra c), LedgerSupportsProtocol (ShelleyBlock (TPraos c) (AlonzoEra c)), ShelleyCompatible (Praos c) (BabbageEra c), LedgerSupportsProtocol (ShelleyBlock (Praos c) (BabbageEra c)), ShelleyCompatible (Praos c) (ConwayEra c), LedgerSupportsProtocol (ShelleyBlock (Praos c) (ConwayEra c)), HASH c ~ Blake2b_256, ADDRHASH c ~ Blake2b_224, DSIGN c ~ Ed25519DSIGN)

newtype CardanoHardForkTriggers

Constructors

CardanoHardForkTriggers

Fields

getCardanoHardForkTriggers :: NP (K TriggerHardFork) (CardanoShelleyEras StandardCrypto)

Bundled Patterns

pattern CardanoHardForkTriggers' :: TriggerHardFork -> TriggerHardFork -> TriggerHardFork -> TriggerHardFork -> TriggerHardFork -> TriggerHardFork -> CardanoHardForkTriggers

type CardanoProtocolParams c = ProtocolParams (CardanoBlock c)

newtype MaxMajorProtVer

Constructors

MaxMajorProtVer

Fields

getMaxMajorProtVer :: Version

data family ProtocolParams blk

data TriggerHardFork

Constructors

TriggerHardForkAtVersion !Word16

TriggerHardForkAtEpoch !EpochNo

TriggerHardForkNotDuringThisExecution

protocolClientInfoCardano :: forall c. EpochSlots -> ProtocolClientInfo (CardanoBlock c)

protocolInfoCardano :: forall c m. (IOLike m, CardanoHardForkConstraints c) => CardanoProtocolParams c -> (ProtocolInfo (CardanoBlock c), m [BlockForging m (CardanoBlock c)])

Create a ProtocolInfo for CardanoBlock

NOTE: For testing and benchmarking purposes, the ShelleyGenesis can contain initial staking and funds. These are registered in the initial ledger state only if the given CardanoHardForkTriggers tell us to skip the Byron era and hard fork directly to Shelley or a later era by using TestXHardForkAtEpoch 0. When gNetworkId == Mainnet, the initial staking and funds must be empty.

PRECONDITION: only a single set of Shelley credentials is allowed when used for mainnet (check against gNetworkId == Mainnet).

SupportedNetworkProtocolVersion

pattern CardanoNodeToClientVersion1 :: BlockNodeToClientVersion (CardanoBlock c)

We support the sole Byron version with the hard fork disabled.

pattern CardanoNodeToClientVersion10 :: BlockNodeToClientVersion (CardanoBlock c)

The hard fork enabled, and the Shelley, Allegra, Mary, Alonzo and Babbage eras enabled Using ShelleyNodeToClientVersion6 for the Shelley-based eras, which enables new queries.

pattern CardanoNodeToClientVersion11 :: BlockNodeToClientVersion (CardanoBlock c)

The hard fork enabled, and the Shelley, Allegra, Mary, Alonzo and Babbage eras enabled, using ShelleyNodeToClientVersion7 for the Shelley-based eras, which enables new queries.

pattern CardanoNodeToClientVersion12 :: BlockNodeToClientVersion (CardanoBlock c)

The hard fork enabled, and the Shelley, Allegra, Mary, Alonzo and Babbage and Conway eras enabled, using ShelleyNodeToClientVersion7 for the Shelley-based eras.

pattern CardanoNodeToClientVersion2 :: BlockNodeToClientVersion (CardanoBlock c)

The hard fork enabled and the Shelley era enabled.

pattern CardanoNodeToClientVersion3 :: BlockNodeToClientVersion (CardanoBlock c)

The hard fork enabled and the Shelley era enabled, but using ShelleyNodeToClientVersion2 and HardForkSpecificNodeToClientVersion2.

pattern CardanoNodeToClientVersion4 :: BlockNodeToClientVersion (CardanoBlock c)

The hard fork enabled, and the Shelley and Allegra eras enabled.

We don't bother with ShelleyNodeToClientVersion1 and HardForkSpecificNodeToClientVersion1.

pattern CardanoNodeToClientVersion5 :: BlockNodeToClientVersion (CardanoBlock c)

The hard fork enabled, and the Shelley, Allegra, and Mary eras enabled.

We don't bother with ShelleyNodeToClientVersion1.

pattern CardanoNodeToClientVersion6 :: BlockNodeToClientVersion (CardanoBlock c)

The hard fork enabled, and the Shelley, Allegra, and Mary eras enabled, but using ShelleyNodeToClientVersion3 for the Shelley-based eras , which enables new queries.

pattern CardanoNodeToClientVersion7 :: BlockNodeToClientVersion (CardanoBlock c)

The hard fork enabled, and the Shelley, Allegra, Mary and Alonzo eras enabled

pattern CardanoNodeToClientVersion8 :: BlockNodeToClientVersion (CardanoBlock c)

The hard fork enabled, and the Shelley, Allegra, Mary and Alonzo eras enabled Using ShelleyNodeToClientVersion5 for the Shelley-based eras , which enables new queries.

pattern CardanoNodeToClientVersion9 :: BlockNodeToClientVersion (CardanoBlock c)

The hard fork enabled, and the Shelley, Allegra, Mary, Alonzo and Babbage eras enabled Using ShelleyNodeToClientVersion5 for the Shelley-based eras, which enables new queries.

pattern CardanoNodeToNodeVersion1 :: BlockNodeToNodeVersion (CardanoBlock c)

We support only Byron V1 with the hard fork disabled, as no other versions have been released before the hard fork

pattern CardanoNodeToNodeVersion2 :: BlockNodeToNodeVersion (CardanoBlock c)

The hard fork enabled using the latest version of Byron and Shelley for each Byron and Shelley era.

Orphan instances

SerialiseConstraintsHFC ByronBlock

Instance details

CardanoHardForkConstraints c => SerialiseHFC (CardanoEras c)
Important: we need to maintain binary compatibility with Byron blocks, as they are already stored on disk.

We also want to be able to efficiently detect (without having to peek far ahead) whether we're dealing with a Byron or Shelley block, so that we can invoke the right decoder. We plan to have a few more hard forks after Shelley (Goguen, Basho, Voltaire), so we want a future-proof envelope for distinguishing the different block types, i.e., a byte indicating the era.

Byron does not provide such an envelope. However, a Byron block is a CBOR 2-tuple with the first element being a tag (Word: 0 = EBB; 1 = regular block) and the second being the payload. We can easily extend this encoding format with support for Shelley, Goguen, etc.

We encode a CardanoBlock as the same CBOR 2-tuple as a Byron block, but we use the tags after 1 for the hard forks after Byron:

Byron EBB

Byron regular block

Shelley block

Allegra block

Mary block

Goguen block

etc.

For more details, see: https://github.com/IntersectMBO/ouroboros-network/pull/1175#issuecomment-558147194

Instance details

Methods

encodeDiskHfcBlock :: CodecConfig (HardForkBlock (CardanoEras c)) -> HardForkBlock (CardanoEras c) -> Encoding

decodeDiskHfcBlock :: CodecConfig (HardForkBlock (CardanoEras c)) -> forall s. Decoder s (ByteString -> HardForkBlock (CardanoEras c))

reconstructHfcPrefixLen :: proxy (Header (HardForkBlock (CardanoEras c))) -> PrefixLen

reconstructHfcNestedCtxt :: proxy (Header (HardForkBlock (CardanoEras c))) -> ShortByteString -> SizeInBytes -> SomeSecond (NestedCtxt Header) (HardForkBlock (CardanoEras c))

getHfcBinaryBlockInfo :: HardForkBlock (CardanoEras c) -> BinaryBlockInfo

estimateHfcBlockSize :: Header (HardForkBlock (CardanoEras c)) -> SizeInBytes

CardanoHardForkConstraints c => SupportedNetworkProtocolVersion (CardanoBlock c)

Instance details

Methods

supportedNodeToNodeVersions :: Proxy (CardanoBlock c) -> Map NodeToNodeVersion (BlockNodeToNodeVersion (CardanoBlock c))

supportedNodeToClientVersions :: Proxy (CardanoBlock c) -> Map NodeToClientVersion (BlockNodeToClientVersion (CardanoBlock c))

latestReleasedNodeVersion :: Proxy (CardanoBlock c) -> (Maybe NodeToNodeVersion, Maybe NodeToClientVersion)