Ouroboros.Consensus.Cardano.ShelleyBased

Injection from Shelley-based eras into the Cardano eras

class cardanoEra ~ ShelleyBlock shelleyEra ⇒ InjectShelley shelleyEra cardanoEra

Witness the relation between the Cardano eras and the Shelley-based eras.

injectShelleyNP ∷ AllZip InjectShelley shelleyEras cardanoEras ⇒ (∀ shelleyEra cardanoEra. InjectShelley shelleyEra cardanoEra ⇒ f shelleyEra → g cardanoEra) → NP f shelleyEras → NP g cardanoEras

injectShelleyOptNP ∷ AllZip InjectShelley shelleyEras cardanoEras ⇒ (∀ shelleyEra cardanoEra. InjectShelley shelleyEra cardanoEra ⇒ f shelleyEra → g cardanoEra) → OptNP empty f shelleyEras → OptNP empty g cardanoEras

Transform Shelley-based types

class EraCrypto era ~ c ⇒ HasCrypto c era

Witness the relation between the crypto used by a Shelley-based era.

Can be partially applied while an equality constraint cannot.

overShelleyBasedLedgerState ∷ ∀ c. PraosCrypto c ⇒ (∀ era. (EraCrypto era ~ c, ShelleyBasedEra era) ⇒ LedgerState (ShelleyBlock era) → LedgerState (ShelleyBlock era)) → LedgerState (CardanoBlock c) → LedgerState (CardanoBlock c)

When the given ledger state corresponds to a Shelley-based era, apply the given function to it.