Ouroboros.Consensus.Shelley.HFEras

Description

Hard fork eras.

Compare this to Eras, which defines ledger eras. This module defines hard fork eras, which are a combination of a ledger era and a protocol.

Documentation

type StandardAllegraBlock = ShelleyBlock (TPraos StandardCrypto) StandardAllegra

type StandardAlonzoBlock = ShelleyBlock (TPraos StandardCrypto) StandardAlonzo

type StandardBabbageBlock = ShelleyBlock (Praos StandardCrypto) StandardBabbage

type StandardConwayBlock = ShelleyBlock (Praos StandardCrypto) StandardConway

type StandardMaryBlock = ShelleyBlock (TPraos StandardCrypto) StandardMary

type StandardShelleyBlock = ShelleyBlock (TPraos StandardCrypto) StandardShelley

Orphan instances

PraosCrypto c => ShelleyCompatible (Praos c) (BabbageEra c)

Instance details

PraosCrypto c => ShelleyCompatible (Praos c) (ConwayEra c)

Instance details

(PraosCrypto c, Signable (DSIGN c) (Hash (HASH c) EraIndependentTxBody)) => ShelleyCompatible (TPraos c) (AllegraEra c)

Instance details

(PraosCrypto c, Signable (DSIGN c) (Hash (HASH c) EraIndependentTxBody)) => ShelleyCompatible (TPraos c) (AlonzoEra c)

Instance details

(PraosCrypto c, PraosCrypto c) => ShelleyCompatible (TPraos c) (BabbageEra c)

Instance details

(PraosCrypto c, PraosCrypto c) => ShelleyCompatible (TPraos c) (ConwayEra c)

Instance details

(PraosCrypto c, Signable (DSIGN c) (Hash (HASH c) EraIndependentTxBody)) => ShelleyCompatible (TPraos c) (MaryEra c)

Instance details

(PraosCrypto c, Signable (DSIGN c) (Hash (HASH c) EraIndependentTxBody)) => ShelleyCompatible (TPraos c) (ShelleyEra c)

Instance details