Ouroboros.Consensus.Shelley.Ledger

Documentation

data family Header blk

data family BlockConfig blk

data family CodecConfig blk

data family StorageConfig blk

data family BlockQuery blk :: Type -> Type

data family LedgerState blk

data family Ticked st

data family GenTx blk

data family TxId tx

data family Validated x

data family NestedCtxt_ blk :: (Type -> Type) -> Type -> Type

data ShelleyBlock proto era

Shelley-based block type.

This block is parametrised over both the (ledger) era and the protocol.

Constructors

ShelleyBlock

Fields

shelleyBlockRaw :: !(Block (ShelleyProtocolHeader proto) era)

shelleyBlockHeaderHash :: !(ShelleyHash (ProtoCrypto proto))

class (ShelleyBasedEra era, ShelleyProtocol proto, Eq (ShelleyProtocolHeader proto), Show (ShelleyProtocolHeader proto), NoThunks (ShelleyProtocolHeader proto), EncCBOR (ShelleyProtocolHeader proto), DecCBOR (Annotator (ShelleyProtocolHeader proto)), Show (CannotForgeError proto), SelectView proto ~ PraosChainSelectView (EraCrypto era), SignedHeader (ShelleyProtocolHeader proto), DecodeDisk (ShelleyBlock proto era) (ChainDepState proto), EncodeDisk (ShelleyBlock proto era) (ChainDepState proto), EraCrypto era ~ ProtoCrypto proto, HasPartialConsensusConfig proto, DecCBOR (PState era), FromCBOR (LegacyPParams era), ToCBOR (LegacyPParams era)) => ShelleyCompatible proto era

class (EraSegWits era, EraGov era, ApplyTx era, ApplyBlock era, EraTransition era, GetLedgerView era, NoThunks (StashedAVVMAddresses era), EncCBOR (StashedAVVMAddresses era), DecCBOR (StashedAVVMAddresses era), Show (StashedAVVMAddresses era), Eq (StashedAVVMAddresses era), DecCBOR (PredicateFailure (EraRule "LEDGER" era)), EncCBOR (PredicateFailure (EraRule "LEDGER" era)), DecCBOR (PredicateFailure (EraRule "UTXOW" era)), EncCBOR (PredicateFailure (EraRule "UTXOW" era)), DSignable (EraCrypto era) (Hash (EraCrypto era) EraIndependentTxBody), NoThunks (PredicateFailure (EraRule "BBODY" era)), NoThunks (TranslationContext era)) => ShelleyBasedEra era

Consensus often needs some more functionality than the ledger currently provides.

Either the functionality shouldn't or can't live in the ledger, in which case it can be part and remain part of ShelleyBasedEra. Or, the functionality should live in the ledger, but hasn't yet been added to the ledger, or it hasn't yet been propagated to this repository, in which case it can be added to this class until that is the case.

If this class becomes redundant, We can move it to ledger and re-export it from here.

TODO Currently we include some constraints on the update state which are needed to determine the hard fork point. In the future this should be replaced with an appropriate API - see https://github.com/IntersectMBO/ouroboros-network/issues/2890

Minimal complete definition

shelleyBasedEraName, applyShelleyBasedTx, getProposedProtocolVersion, getConwayEraGovDict

class HasHeader (Header blk) => GetHeader blk where

Methods

getHeader :: blk -> Header blk

blockMatchesHeader :: Header blk -> blk -> Bool

headerIsEBB :: Header blk -> Maybe EpochNo

newtype ShelleyHash crypto

Constructors

ShelleyHash

Fields

unShelleyHash :: Hash crypto EraIndependentBlockHeader

data CompactGenesis c

Compact variant of ShelleyGenesis with some fields erased that are only used on start-up and that should not be kept in memory forever.

Concretely:

The sgInitialFunds field is erased. It is only used to set up the initial UTxO in tests and testnets.

The sgStaking field is erased. It is only used to register initial stake pools in tests and benchmarks.

newtype ShelleyLedgerError era

Constructors

BBodyError (BlockTransitionError era)

data ShelleyTip proto era

Constructors

ShelleyTip

Fields

shelleyTipSlotNo :: !SlotNo

shelleyTipBlockNo :: !BlockNo

shelleyTipHash :: !(HeaderHash (ShelleyBlock proto era))

newtype ShelleyTransition

Information required to determine the hard fork point from Shelley to the next ledger

Constructors

ShelleyTransitionInfo

Fields

shelleyAfterVoting :: Word32

The number of blocks in this epoch past the voting deadline

We record this to make sure that we can tell the HFC about hard forks if and only if we are certain:

Blocks that came in within an epoch after the 4k/f voting deadline are not relevant (10kf - 2 * 3kf).

Since there are slots between blocks, we are probably only sure that there will be no more relevant block when we have seen the first block after the deadline.

If we count how many blocks we have seen post deadline, and we have reached k of them, we know that that last pre-deadline block won't be rolled back anymore.

At this point we can look at the ledger state and see which proposals we accepted in the voting period, if any, and notify the HFC is one of them indicates a transition.

data ShelleyLedgerConfig era

Constructors

ShelleyLedgerConfig

Fields

shelleyLedgerCompactGenesis :: !(CompactGenesis (EraCrypto era))

shelleyLedgerGlobals :: !Globals

Derived from shelleyLedgerGenesis but we store a cached version because it used very often.

shelleyLedgerTranslationContext :: !(TranslationContext era)

data ShelleyLedgerEvent era

All events emitted by the Shelley ledger API

Constructors

ShelleyLedgerEventBBODY (Event (EraRule "BBODY" era))
An event emitted when (re)applying a block

ShelleyLedgerEventTICK (Event (EraRule "TICK" era))
An event emitted during the chain tick

data ShelleyReapplyException

Constructors

forall era.Show (BlockTransitionError era) => ShelleyReapplyException (BlockTransitionError era)

newtype ApplyTxError era

Constructors

ApplyTxError [PredicateFailure (EraRule "LEDGER" era)]

data WithTop a

Add a unique top element to a lattice.

TODO This should be relocated to `cardano-base:Data.Measure'.

Constructors

NotTop a

Top

data AlonzoMeasure

Constructors

AlonzoMeasure

Fields

byteSize :: !ByteSize

exUnits :: !(ExUnits' (WithTop Natural))

data ShelleyNodeToClientVersion

Constructors

ShelleyNodeToClientVersion1

ShelleyNodeToClientVersion2
New queries introduced

ShelleyNodeToClientVersion3
New query introduced

ShelleyNodeToClientVersion4
New queries introduced

ShelleyNodeToClientVersion5
New queries introduced: GetRewardInfoPools

ShelleyNodeToClientVersion6
New queries introduced: GetPoolDistr, GetPoolState, GetStakeSnapshots

ShelleyNodeToClientVersion7
New queries introduced: GetStakeDelegDeposits

ShelleyNodeToClientVersion8
New queries introduced: GetConstitutionHash, GetFilteredVoteDelegatees

data ShelleyNodeToNodeVersion

Constructors

ShelleyNodeToNodeVersion1

newtype NonMyopicMemberRewards c

Constructors

NonMyopicMemberRewards

Fields

unNonMyopicMemberRewards :: Map (Either Coin (Credential 'Staking c)) (Map (KeyHash 'StakePool c) Coin)

data StakeSnapshot crypto

The stake snapshot returns information about the mark, set, go ledger snapshots for a pool, plus the total active stake for each snapshot that can be used in a sigma calculation.

Each snapshot is taken at the end of a different era. The go snapshot is the current one and was taken two epochs earlier, set was taken one epoch ago, and mark was taken immediately before the start of the current epoch.

Constructors

StakeSnapshot

Fields

ssMarkPool :: !Coin

ssSetPool :: !Coin

ssGoPool :: !Coin

data StakeSnapshots crypto

Constructors

StakeSnapshots

Fields

ssStakeSnapshots :: !(Map (KeyHash 'StakePool crypto) (StakeSnapshot crypto))

ssMarkTotal :: !Coin

ssSetTotal :: !Coin

ssGoTotal :: !Coin

verifyBlockIntegrity :: ShelleyCompatible proto era => Word64 -> ShelleyBlock proto era -> Bool

Verifies whether the block is not corrupted by checking its signature and witnesses.

verifyHeaderIntegrity

Arguments

:: ProtocolHeaderSupportsKES proto

=> Word64
Slots per KES period

-> ShelleyProtocolHeader proto

-> Bool

Verify that the signature on a header is correct and valid.

mkShelleyBlock :: ShelleyCompatible proto era => Block (ShelleyProtocolHeader proto) era -> ShelleyBlock proto era

mkShelleyHeader :: ShelleyCompatible proto era => ShelleyProtocolHeader proto -> Header (ShelleyBlock proto era)

decodeShelleyBlock :: forall proto era. ShelleyCompatible proto era => forall s. Decoder s (ByteString -> ShelleyBlock proto era)

decodeShelleyHeader :: forall proto era. ShelleyCompatible proto era => forall s. Decoder s (ByteString -> Header (ShelleyBlock proto era))

encodeShelleyBlock :: forall proto era. ShelleyCompatible proto era => ShelleyBlock proto era -> Encoding

encodeShelleyHeader :: forall proto era. ShelleyCompatible proto era => Header (ShelleyBlock proto era) -> Encoding

shelleyBinaryBlockInfo :: forall proto era. ShelleyCompatible proto era => ShelleyBlock proto era -> BinaryBlockInfo

fromShelleyPrevHash :: EraCrypto era ~ ProtoCrypto proto => PrevHash (EraCrypto era) -> ChainHash (ShelleyBlock proto era)

From cardano-ledger-specs to ouroboros-consensus

toShelleyPrevHash :: EraCrypto era ~ ProtoCrypto proto => ChainHash (Header (ShelleyBlock proto era)) -> PrevHash (EraCrypto era)

From ouroboros-consensus to cardano-ledger-specs

compactGenesis :: ShelleyGenesis c -> CompactGenesis c

Compacts the given ShelleyGenesis.

mkShelleyBlockConfig :: ShelleyBasedEra era => ProtVer -> ShelleyGenesis (EraCrypto era) -> [VKey 'BlockIssuer (EraCrypto era)] -> BlockConfig (ShelleyBlock proto era)

forgeShelleyBlock

Arguments

:: forall m era proto. (ShelleyCompatible proto era, TxLimits (ShelleyBlock proto era), Monad m)

=> HotKey (EraCrypto era) m

-> CanBeLeader proto

-> TopLevelConfig (ShelleyBlock proto era)

-> TxOverrides (ShelleyBlock proto era)
How to override max tx capacity defined by ledger

-> BlockNo
Current block number

-> SlotNo
Current slot number

-> TickedLedgerState (ShelleyBlock proto era)
Current ledger

-> [Validated (GenTx (ShelleyBlock proto era))]
Txs to add in the block

-> IsLeader proto

-> m (ShelleyBlock proto era)

castShelleyTip :: HeaderHash (ShelleyBlock proto era) ~ HeaderHash (ShelleyBlock proto' era') => ShelleyTip proto era -> ShelleyTip proto' era'

shelleyLedgerTipPoint :: LedgerState (ShelleyBlock proto era) -> Point (ShelleyBlock proto era)

shelleyTipToPoint :: WithOrigin (ShelleyTip proto era) -> Point (ShelleyBlock proto era)

mkShelleyLedgerConfig :: ShelleyGenesis (EraCrypto era) -> TranslationContext era -> EpochInfo (Except PastHorizonException) -> MaxMajorProtVer -> ShelleyLedgerConfig era

shelleyEraParams :: ShelleyGenesis c -> EraParams

shelleyEraParamsNeverHardForks :: ShelleyGenesis c -> EraParams

Separate variant of shelleyEraParams to be used for a Shelley-only chain.

shelleyLedgerGenesis :: ShelleyLedgerConfig era -> ShelleyGenesis (EraCrypto era)

getPParams :: EraGov era => NewEpochState era -> PParams era

decodeShelleyAnnTip :: ShelleyCompatible proto era => Decoder s (AnnTip (ShelleyBlock proto era))

decodeShelleyLedgerState :: forall era proto s. ShelleyCompatible proto era => Decoder s (LedgerState (ShelleyBlock proto era))

encodeShelleyAnnTip :: ShelleyCompatible proto era => AnnTip (ShelleyBlock proto era) -> Encoding

encodeShelleyHeaderState :: ShelleyCompatible proto era => HeaderState (ShelleyBlock proto era) -> Encoding

encodeShelleyLedgerState :: ShelleyCompatible proto era => LedgerState (ShelleyBlock proto era) -> Encoding

fixedBlockBodyOverhead :: Num a => a

txInBlockSize is used to estimate how many transactions we can grab from the Mempool to put into the block we are going to forge without exceeding the maximum block body size according to the ledger. If we exceed that limit, we will have forged a block that is invalid according to the ledger. We ourselves won't even adopt it, causing us to lose our slot, something we must try to avoid.

For this reason it is better to overestimate the size of a transaction than to underestimate. The only downside is that we maybe could have put one (or more?) transactions extra in that block.

As the sum of the serialised transaction sizes is not equal to the size of the serialised block body (TxSeq) consisting of those transactions (see cardano-node#1545 for an example), we account for some extra overhead per transaction as a safety margin.

Also see perTxOverhead.

mkShelleyTx :: forall era proto. ShelleyBasedEra era => Tx era -> GenTx (ShelleyBlock proto era)

mkShelleyValidatedTx :: forall era proto. ShelleyBasedEra era => Validated (Tx era) -> Validated (GenTx (ShelleyBlock proto era))

perTxOverhead :: Num a => a

See fixedBlockBodyOverhead.

fromExUnits :: ExUnits -> ExUnits' (WithTop Natural)

querySupportedVersion :: BlockQuery (ShelleyBlock proto era) result -> ShelleyNodeToClientVersion -> Bool

Is the given query supported by the given ShelleyNodeToClientVersion?

decodeShelleyQuery :: forall era proto. ShelleyBasedEra era => forall s. Decoder s (SomeSecond BlockQuery (ShelleyBlock proto era))

decodeShelleyResult :: forall proto era result. ShelleyCompatible proto era => ShelleyNodeToClientVersion -> BlockQuery (ShelleyBlock proto era) result -> forall s. Decoder s result

encodeShelleyQuery :: forall era proto result. ShelleyBasedEra era => BlockQuery (ShelleyBlock proto era) result -> Encoding

encodeShelleyResult :: forall proto era result. ShelleyCompatible proto era => ShelleyNodeToClientVersion -> BlockQuery (ShelleyBlock proto era) result -> result -> Encoding