Ouroboros.Consensus.Shelley.Ledger.Inspect

Documentation

data ProtocolUpdate era

Constructors

ProtocolUpdate

Fields

protocolUpdateProposal :: UpdateProposal era

protocolUpdateState :: UpdateState (EraCrypto era)

data ShelleyLedgerUpdate era

Constructors

ShelleyUpdatedProtocolUpdates [ProtocolUpdate era]

data UpdateProposal era

Update proposal

As in Byron, a proposal is a partial map from parameters to their values.

Constructors

UpdateProposal

Fields

proposalParams :: PParamsUpdate era

The protocol parameters changed by this update proposal

An update is identified by how it updates the protocol parameters.

proposalVersion :: Maybe ProtVer

New version (if changed by this proposal)

The protocol version itself is also considered to be just another parameter, and parameters can change without changing the protocol version, although a convention could be established that the protocol version must change if any of the parameters do; but the specification itself does not mandate this.

We record the version separately for the convenience of the HFC.

proposalEpoch :: EpochNo

The EpochNo the proposal becomes active in, if it is adopted

data UpdateState c

Proposal state

The update mechanism in Shelley is simpler than it is in Byron. There is no distinction between votes and proposals: to "vote" for a proposal one merely submits the exact same proposal. There is also no separate endorsement step. The procedure is as follows:

During each epoch, a genesis key can submit (via its delegates) zero, one, or many proposals; each submission overrides the previous one.

"Voting" (submitting of proposals) ends 2 * stabilityWindow slots (i.e. 6k/f) before the end of the epoch. In other words, proposals for the upcoming epoch must be submitted within the first 4k/f slots of this one.

At the end of an epoch, if the majority of nodes (as determined by the Quorum specification constant, which must be greater than half the nodes) have most recently submitted the same exact proposal, then it is adopted.

The next epoch is always started with a clean slate, proposals from the previous epoch that didn't make it are discarded (except for "future proposals" that are explicitly marked for future epochs).

Constructors

UpdateState

Fields

proposalVotes :: [KeyHash 'Genesis c]

The genesis delegates that voted for this proposal

proposalReachedQuorum :: Bool

Has this proposal reached sufficient votes to be adopted?

protocolUpdates :: forall era proto. ShelleyBasedEra era => ShelleyGenesis (EraCrypto era) -> LedgerState (ShelleyBlock proto era) -> [ProtocolUpdate era]

Orphan instances

ShelleyBasedEra era => InspectLedger (ShelleyBlock proto era)

Instance details

Associated Types

type LedgerWarning (ShelleyBlock proto era)

type LedgerUpdate (ShelleyBlock proto era)

Methods

inspectLedger :: TopLevelConfig (ShelleyBlock proto era) -> LedgerState (ShelleyBlock proto era) -> LedgerState (ShelleyBlock proto era) -> [LedgerEvent (ShelleyBlock proto era)]