Ouroboros.Consensus.Shelley.Ledger.Integrity

Documentation

verifyBlockIntegrity :: ShelleyCompatible proto era => Word64 -> ShelleyBlock proto era -> Bool

Verifies whether the block is not corrupted by checking its signature and witnesses.

verifyHeaderIntegrity

Arguments

:: ProtocolHeaderSupportsKES proto

=> Word64
Slots per KES period

-> ShelleyProtocolHeader proto

-> Bool

Verify that the signature on a header is correct and valid.