Ouroboros.Consensus.Shelley.Ledger.Ledger

Documentation

data family LedgerState blk

class (EraSegWits era, EraGov era, ApplyTx era, ApplyBlock era, EraTransition era, GetLedgerView era, NoThunks (StashedAVVMAddresses era), EncCBOR (StashedAVVMAddresses era), DecCBOR (StashedAVVMAddresses era), Show (StashedAVVMAddresses era), Eq (StashedAVVMAddresses era), DecCBOR (PredicateFailure (EraRule "LEDGER" era)), EncCBOR (PredicateFailure (EraRule "LEDGER" era)), DecCBOR (PredicateFailure (EraRule "UTXOW" era)), EncCBOR (PredicateFailure (EraRule "UTXOW" era)), DSignable (EraCrypto era) (Hash (EraCrypto era) EraIndependentTxBody), NoThunks (PredicateFailure (EraRule "BBODY" era)), NoThunks (TranslationContext era)) => ShelleyBasedEra era

Consensus often needs some more functionality than the ledger currently provides.

Either the functionality shouldn't or can't live in the ledger, in which case it can be part and remain part of ShelleyBasedEra. Or, the functionality should live in the ledger, but hasn't yet been added to the ledger, or it hasn't yet been propagated to this repository, in which case it can be added to this class until that is the case.

If this class becomes redundant, We can move it to ledger and re-export it from here.

TODO Currently we include some constraints on the update state which are needed to determine the hard fork point. In the future this should be replaced with an appropriate API - see https://github.com/IntersectMBO/ouroboros-network/issues/2890

Minimal complete definition

shelleyBasedEraName, applyShelleyBasedTx, getProposedProtocolVersion, getConwayEraGovDict

newtype ShelleyLedgerError era

Constructors

BBodyError (BlockTransitionError era)

data ShelleyTip proto era

Constructors

ShelleyTip

Fields

shelleyTipSlotNo :: !SlotNo

shelleyTipBlockNo :: !BlockNo

shelleyTipHash :: !(HeaderHash (ShelleyBlock proto era))

newtype ShelleyTransition

Information required to determine the hard fork point from Shelley to the next ledger

Constructors

ShelleyTransitionInfo

Fields

shelleyAfterVoting :: Word32

The number of blocks in this epoch past the voting deadline

We record this to make sure that we can tell the HFC about hard forks if and only if we are certain:

Blocks that came in within an epoch after the 4k/f voting deadline are not relevant (10kf - 2 * 3kf).

Since there are slots between blocks, we are probably only sure that there will be no more relevant block when we have seen the first block after the deadline.

If we count how many blocks we have seen post deadline, and we have reached k of them, we know that that last pre-deadline block won't be rolled back anymore.

At this point we can look at the ledger state and see which proposals we accepted in the voting period, if any, and notify the HFC is one of them indicates a transition.

data family Ticked st

castShelleyTip :: HeaderHash (ShelleyBlock proto era) ~ HeaderHash (ShelleyBlock proto' era') => ShelleyTip proto era -> ShelleyTip proto' era'

shelleyLedgerTipPoint :: LedgerState (ShelleyBlock proto era) -> Point (ShelleyBlock proto era)

shelleyTipToPoint :: WithOrigin (ShelleyTip proto era) -> Point (ShelleyBlock proto era)

Ledger config

data ShelleyLedgerConfig era

Constructors

ShelleyLedgerConfig

Fields

shelleyLedgerCompactGenesis :: !(CompactGenesis (EraCrypto era))

shelleyLedgerGlobals :: !Globals

Derived from shelleyLedgerGenesis but we store a cached version because it used very often.

shelleyLedgerTranslationContext :: !(TranslationContext era)

mkShelleyLedgerConfig :: ShelleyGenesis (EraCrypto era) -> TranslationContext era -> EpochInfo (Except PastHorizonException) -> MaxMajorProtVer -> ShelleyLedgerConfig era

shelleyEraParams :: ShelleyGenesis c -> EraParams

shelleyEraParamsNeverHardForks :: ShelleyGenesis c -> EraParams

Separate variant of shelleyEraParams to be used for a Shelley-only chain.

shelleyLedgerGenesis :: ShelleyLedgerConfig era -> ShelleyGenesis (EraCrypto era)

Auxiliary

data ShelleyLedgerEvent era

All events emitted by the Shelley ledger API

Constructors

ShelleyLedgerEventBBODY (Event (EraRule "BBODY" era))
An event emitted when (re)applying a block

ShelleyLedgerEventTICK (Event (EraRule "TICK" era))
An event emitted during the chain tick

data ShelleyReapplyException

Constructors

forall era.Show (BlockTransitionError era) => ShelleyReapplyException (BlockTransitionError era)

getPParams :: EraGov era => NewEpochState era -> PParams era

Serialisation

decodeShelleyAnnTip :: ShelleyCompatible proto era => Decoder s (AnnTip (ShelleyBlock proto era))

decodeShelleyLedgerState :: forall era proto s. ShelleyCompatible proto era => Decoder s (LedgerState (ShelleyBlock proto era))

encodeShelleyAnnTip :: ShelleyCompatible proto era => AnnTip (ShelleyBlock proto era) -> Encoding

encodeShelleyHeaderState :: ShelleyCompatible proto era => HeaderState (ShelleyBlock proto era) -> Encoding

encodeShelleyLedgerState :: ShelleyCompatible proto era => LedgerState (ShelleyBlock proto era) -> Encoding

Orphan instances

Generic (LedgerState (ShelleyBlock proto era))

Instance details

Associated Types

type Rep (LedgerState (ShelleyBlock proto era)) :: Type -> Type

Methods

from :: LedgerState (ShelleyBlock proto era) -> Rep (LedgerState (ShelleyBlock proto era)) x

to :: Rep (LedgerState (ShelleyBlock proto era)) x -> LedgerState (ShelleyBlock proto era)

Generic (Ticked (LedgerState (ShelleyBlock proto era)))

Instance details

Associated Types

type Rep (Ticked (LedgerState (ShelleyBlock proto era))) :: Type -> Type

Methods

from :: Ticked (LedgerState (ShelleyBlock proto era)) -> Rep (Ticked (LedgerState (ShelleyBlock proto era))) x

to :: Rep (Ticked (LedgerState (ShelleyBlock proto era))) x -> Ticked (LedgerState (ShelleyBlock proto era))

ShelleyBasedEra era => Show (LedgerState (ShelleyBlock proto era))

Instance details

Methods

showsPrec :: Int -> LedgerState (ShelleyBlock proto era) -> ShowS

show :: LedgerState (ShelleyBlock proto era) -> String

showList :: [LedgerState (ShelleyBlock proto era)] -> ShowS

ShelleyBasedEra era => Eq (LedgerState (ShelleyBlock proto era))

Instance details

Methods

(==) :: LedgerState (ShelleyBlock proto era) -> LedgerState (ShelleyBlock proto era) -> Bool

(/=) :: LedgerState (ShelleyBlock proto era) -> LedgerState (ShelleyBlock proto era) -> Bool

ShelleyBasedEra era => NoThunks (LedgerState (ShelleyBlock proto era))

Instance details

Methods

noThunks :: Context -> LedgerState (ShelleyBlock proto era) -> IO (Maybe ThunkInfo)

wNoThunks :: Context -> LedgerState (ShelleyBlock proto era) -> IO (Maybe ThunkInfo)

showTypeOf :: Proxy (LedgerState (ShelleyBlock proto era)) -> String

ShelleyBasedEra era => NoThunks (Ticked (LedgerState (ShelleyBlock proto era)))

Instance details

Methods

noThunks :: Context -> Ticked (LedgerState (ShelleyBlock proto era)) -> IO (Maybe ThunkInfo)

wNoThunks :: Context -> Ticked (LedgerState (ShelleyBlock proto era)) -> IO (Maybe ThunkInfo)

showTypeOf :: Proxy (Ticked (LedgerState (ShelleyBlock proto era))) -> String

GetTip (LedgerState (ShelleyBlock proto era))

Instance details

Methods

getTip :: LedgerState (ShelleyBlock proto era) -> Point (LedgerState (ShelleyBlock proto era))

GetTip (Ticked (LedgerState (ShelleyBlock proto era)))

Instance details

Methods

getTip :: Ticked (LedgerState (ShelleyBlock proto era)) -> Point (Ticked (LedgerState (ShelleyBlock proto era)))

ShelleyBasedEra era => IsLedger (LedgerState (ShelleyBlock proto era))

Instance details

Associated Types

type LedgerErr (LedgerState (ShelleyBlock proto era))

type AuxLedgerEvent (LedgerState (ShelleyBlock proto era))

Methods

applyChainTickLedgerResult :: LedgerCfg (LedgerState (ShelleyBlock proto era)) -> SlotNo -> LedgerState (ShelleyBlock proto era) -> LedgerResult (LedgerState (ShelleyBlock proto era)) (Ticked (LedgerState (ShelleyBlock proto era)))

ShelleyCompatible proto era => ApplyBlock (LedgerState (ShelleyBlock proto era)) (ShelleyBlock proto era)

Instance details

Methods

applyBlockLedgerResult :: LedgerCfg (LedgerState (ShelleyBlock proto era)) -> ShelleyBlock proto era -> Ticked (LedgerState (ShelleyBlock proto era)) -> Except (LedgerErr (LedgerState (ShelleyBlock proto era))) (LedgerResult (LedgerState (ShelleyBlock proto era)) (LedgerState (ShelleyBlock proto era)))

reapplyBlockLedgerResult :: LedgerCfg (LedgerState (ShelleyBlock proto era)) -> ShelleyBlock proto era -> Ticked (LedgerState (ShelleyBlock proto era)) -> LedgerResult (LedgerState (ShelleyBlock proto era)) (LedgerState (ShelleyBlock proto era))

HasHardForkHistory (ShelleyBlock proto era)

Instance details

Associated Types

type HardForkIndices (ShelleyBlock proto era) :: [Type]

Methods

hardForkSummary :: LedgerConfig (ShelleyBlock proto era) -> LedgerState (ShelleyBlock proto era) -> Summary (HardForkIndices (ShelleyBlock proto era))

ShelleyCompatible proto era => BasicEnvelopeValidation (ShelleyBlock proto era)

Instance details

Methods

expectedFirstBlockNo :: proxy (ShelleyBlock proto era) -> BlockNo

expectedNextBlockNo :: proxy (ShelleyBlock proto era) -> TipInfo (ShelleyBlock proto era) -> TipInfo (ShelleyBlock proto era) -> BlockNo -> BlockNo

minimumPossibleSlotNo :: Proxy (ShelleyBlock proto era) -> SlotNo

minimumNextSlotNo :: proxy (ShelleyBlock proto era) -> TipInfo (ShelleyBlock proto era) -> TipInfo (ShelleyBlock proto era) -> SlotNo -> SlotNo

ShelleyCompatible proto era => ValidateEnvelope (ShelleyBlock proto era)

Instance details

Associated Types

type OtherHeaderEnvelopeError (ShelleyBlock proto era)

Methods

additionalEnvelopeChecks :: TopLevelConfig (ShelleyBlock proto era) -> LedgerView (BlockProtocol (ShelleyBlock proto era)) -> Header (ShelleyBlock proto era) -> Except (OtherHeaderEnvelopeError (ShelleyBlock proto era)) ()

ShelleyCompatible proto era => UpdateLedger (ShelleyBlock proto era)

Instance details

ShelleyCompatible proto era => CommonProtocolParams (ShelleyBlock proto era)

Instance details

Methods

maxHeaderSize :: LedgerState (ShelleyBlock proto era) -> Word32

maxTxSize :: LedgerState (ShelleyBlock proto era) -> Word32