Ouroboros.Consensus.Shelley.Ledger.Query

Documentation

data family BlockQuery blk :: Type -> Type

newtype NonMyopicMemberRewards c

Constructors

NonMyopicMemberRewards

Fields

unNonMyopicMemberRewards :: Map (Either Coin (Credential 'Staking c)) (Map (KeyHash 'StakePool c) Coin)

data StakeSnapshot crypto

The stake snapshot returns information about the mark, set, go ledger snapshots for a pool, plus the total active stake for each snapshot that can be used in a sigma calculation.

Each snapshot is taken at the end of a different era. The go snapshot is the current one and was taken two epochs earlier, set was taken one epoch ago, and mark was taken immediately before the start of the current epoch.

Constructors

StakeSnapshot

Fields

ssMarkPool :: !Coin

ssSetPool :: !Coin

ssGoPool :: !Coin

data StakeSnapshots crypto

Constructors

StakeSnapshots

Fields

ssStakeSnapshots :: !(Map (KeyHash 'StakePool crypto) (StakeSnapshot crypto))

ssMarkTotal :: !Coin

ssSetTotal :: !Coin

ssGoTotal :: !Coin

querySupportedVersion :: BlockQuery (ShelleyBlock proto era) result -> ShelleyNodeToClientVersion -> Bool

Is the given query supported by the given ShelleyNodeToClientVersion?

Serialisation

decodeShelleyQuery :: forall era proto. ShelleyBasedEra era => forall s. Decoder s (SomeSecond BlockQuery (ShelleyBlock proto era))

decodeShelleyResult :: forall proto era result. ShelleyCompatible proto era => ShelleyNodeToClientVersion -> BlockQuery (ShelleyBlock proto era) result -> forall s. Decoder s result

encodeShelleyQuery :: forall era proto result. ShelleyBasedEra era => BlockQuery (ShelleyBlock proto era) result -> Encoding

encodeShelleyResult :: forall proto era result. ShelleyCompatible proto era => ShelleyNodeToClientVersion -> BlockQuery (ShelleyBlock proto era) result -> result -> Encoding

Orphan instances

SameDepIndex (BlockQuery (ShelleyBlock proto era))

Instance details

Methods

sameDepIndex :: BlockQuery (ShelleyBlock proto era) a -> BlockQuery (ShelleyBlock proto era) b -> Maybe (a :~: b)

ShelleyCompatible proto era => ShowQuery (BlockQuery (ShelleyBlock proto era))

Instance details

Methods

showResult :: BlockQuery (ShelleyBlock proto era) result -> result -> String

Show (BlockQuery (ShelleyBlock proto era) result)

Instance details

Methods

showsPrec :: Int -> BlockQuery (ShelleyBlock proto era) result -> ShowS

show :: BlockQuery (ShelleyBlock proto era) result -> String

showList :: [BlockQuery (ShelleyBlock proto era) result] -> ShowS

Eq (BlockQuery (ShelleyBlock proto era) result)

Instance details

Methods

(==) :: BlockQuery (ShelleyBlock proto era) result -> BlockQuery (ShelleyBlock proto era) result -> Bool

(/=) :: BlockQuery (ShelleyBlock proto era) result -> BlockQuery (ShelleyBlock proto era) result -> Bool

(ShelleyCompatible proto era, ProtoCrypto proto ~ crypto) => QueryLedger (ShelleyBlock proto era)

Instance details

Methods

answerBlockQuery :: ExtLedgerCfg (ShelleyBlock proto era) -> BlockQuery (ShelleyBlock proto era) result -> ExtLedgerState (ShelleyBlock proto era) -> result

(Typeable era, Typeable proto) => ShowProxy (BlockQuery (ShelleyBlock proto era) :: Type -> Type)

Instance details

Methods

showProxy :: Proxy (BlockQuery (ShelleyBlock proto era)) -> String