Ouroboros.Consensus.Shelley.Ledger.SupportsProtocol

Description

This module contains SupportsProtocol instances tying the ledger and protocol together. Since these instances reference both ledger concerns and protocol concerns, it is the one class where we cannot provide a generic instance for ShelleyBlock.

Orphan instances

(ShelleyCompatible (Praos crypto) era, ShelleyCompatible (TPraos crypto) era, crypto ~ EraCrypto era, TranslateProto (TPraos crypto) (Praos crypto)) => LedgerSupportsProtocol (ShelleyBlock (Praos crypto) era)

Instance details

Methods

protocolLedgerView :: LedgerConfig (ShelleyBlock (Praos crypto) era) -> Ticked (LedgerState (ShelleyBlock (Praos crypto) era)) -> LedgerView (BlockProtocol (ShelleyBlock (Praos crypto) era))

ledgerViewForecastAt :: LedgerConfig (ShelleyBlock (Praos crypto) era) -> LedgerState (ShelleyBlock (Praos crypto) era) -> Forecast (LedgerView (BlockProtocol (ShelleyBlock (Praos crypto) era)))

(ShelleyCompatible (TPraos crypto) era, crypto ~ EraCrypto era) => LedgerSupportsProtocol (ShelleyBlock (TPraos crypto) era)

Instance details

Methods

protocolLedgerView :: LedgerConfig (ShelleyBlock (TPraos crypto) era) -> Ticked (LedgerState (ShelleyBlock (TPraos crypto) era)) -> LedgerView (BlockProtocol (ShelleyBlock (TPraos crypto) era))

ledgerViewForecastAt :: LedgerConfig (ShelleyBlock (TPraos crypto) era) -> LedgerState (ShelleyBlock (TPraos crypto) era) -> Forecast (LedgerView (BlockProtocol (ShelleyBlock (TPraos crypto) era)))