Ouroboros.Consensus.Shelley.Node

Documentation

newtype MaxMajorProtVer

Constructors

MaxMajorProtVer

Fields

getMaxMajorProtVer :: Version

data family ProtocolParams blk

data ProtocolParamsShelleyBased c

Parameters common to all Shelley-based ledgers.

When running a chain with multiple Shelley-based eras, in addition to the per-era protocol parameters, one value of ProtocolParamsShelleyBased will be needed, which is shared among all Shelley-based eras.

Constructors

ProtocolParamsShelleyBased

Fields

shelleyBasedInitialNonce :: Nonce

The initial nonce, typically derived from the hash of Genesis config JSON file.

WARNING: chains using different values of this parameter will be mutually incompatible.

shelleyBasedLeaderCredentials :: [ShelleyLeaderCredentials c]

data Nonce

Constructors

Nonce !(Hash Blake2b_256 Nonce)

NeutralNonce

data ProtVer

Constructors

ProtVer

Fields

pvMajor :: !Version

pvMinor :: !Natural

data ShelleyGenesis c

Constructors

ShelleyGenesis

Fields

sgSystemStart :: !UTCTime

sgNetworkMagic :: !Word32

sgNetworkId :: !Network

sgActiveSlotsCoeff :: !PositiveUnitInterval

sgSecurityParam :: !Word64

sgEpochLength :: !EpochSize

sgSlotsPerKESPeriod :: !Word64

sgMaxKESEvolutions :: !Word64

sgSlotLength :: !NominalDiffTimeMicro

sgUpdateQuorum :: !Word64

sgMaxLovelaceSupply :: !Word64

sgProtocolParams :: !(PParams (ShelleyEra c))

sgGenDelegs :: !(Map (KeyHash 'Genesis c) (GenDelegPair c))

sgInitialFunds :: ListMap (Addr c) Coin

sgStaking :: ShelleyGenesisStaking c

data ShelleyGenesisStaking c

Constructors

ShelleyGenesisStaking

Fields

sgsPools :: ListMap (KeyHash 'StakePool c) (PoolParams c)

sgsStake :: ListMap (KeyHash 'Staking c) (KeyHash 'StakePool c)

emptyGenesisStaking :: ShelleyGenesisStaking c

data ShelleyLeaderCredentials c

Constructors

ShelleyLeaderCredentials

Fields

shelleyLeaderCredentialsInitSignKey :: SignKeyKES c

The unevolved signing KES key (at evolution 0).

Note that this is not inside ShelleyCanBeLeader since it gets evolved automatically, whereas ShelleyCanBeLeader does not change.

shelleyLeaderCredentialsCanBeLeader :: PraosCanBeLeader c

shelleyLeaderCredentialsLabel :: Text

Identifier for this set of credentials.

Useful when the node is running with multiple sets of credentials.

protocolClientInfoShelley :: ProtocolClientInfo (ShelleyBlock proto era)

protocolInfoShelley :: forall m c. (IOLike m, PraosCrypto c, ShelleyCompatible (TPraos c) (ShelleyEra c), TxLimits (ShelleyBlock (TPraos c) (ShelleyEra c))) => ShelleyGenesis c -> ProtocolParamsShelleyBased c -> ProtocolParams (ShelleyBlock (TPraos c) (ShelleyEra c)) -> (ProtocolInfo (ShelleyBlock (TPraos c) (ShelleyEra c)), m [BlockForging m (ShelleyBlock (TPraos c) (ShelleyEra c))])

protocolInfoTPraosShelleyBased

Arguments

:: forall m era c. (IOLike m, PraosCrypto c, ShelleyCompatible (TPraos c) era, TxLimits (ShelleyBlock (TPraos c) era), c ~ EraCrypto era)

=> ProtocolParamsShelleyBased c

-> TransitionConfig era

-> ProtVer
see $sel:shelleyProtVer:ProtocolParamsShelley, mutatis mutandi

-> TxOverrides (ShelleyBlock (TPraos c) era)

-> (ProtocolInfo (ShelleyBlock (TPraos c) era), m [BlockForging m (ShelleyBlock (TPraos c) era)])

validateGenesis :: PraosCrypto c => ShelleyGenesis c -> Either String ()

Check the validity of the genesis config. To be used in conjunction with assertWithMsg.

Orphan instances

ShelleyCompatible proto era => BlockSupportsMetrics (ShelleyBlock proto era)

Instance details

Methods

isSelfIssued :: BlockConfig (ShelleyBlock proto era) -> Header (ShelleyBlock proto era) -> WhetherSelfIssued

(ShelleyCompatible proto era, LedgerSupportsProtocol (ShelleyBlock proto era)) => RunNode (ShelleyBlock proto era)

Instance details