Ouroboros.Consensus.Cardano

The block type of the Cardano block chain

type CardanoBlock c = HardForkBlock (CardanoEras c)

The Cardano block.

Thanks to the pattern synonyms, you can treat this as a sum type with constructors BlockByron and BlockShelley.

toEither :: CardanoBlock c -> Either ByronBlock (ShelleyBlock (ShelleyEra c))
toEither (BlockByron   b) = Left b
toEither (BlockShelley s) = Right s

Supported protocols

type ProtocolByron = HardForkProtocol '[ByronBlock]

type ProtocolShelley = HardForkProtocol '[ShelleyBlock StandardShelley]

type ProtocolCardano = HardForkProtocol '[ByronBlock, ShelleyBlock StandardShelley]

Abstract over the various protocols

data Protocol (m ∷ Type → Type) blk p where

Consensus protocol to use

Constructors

ProtocolByron ∷ Config → Maybe PBftSignatureThreshold → ProtocolVersion → SoftwareVersion → Maybe ByronLeaderCredentials → Protocol m ByronBlockHFC ProtocolByron
Run PBFT against the real Byron ledger

ProtocolShelley
Run TPraos against the real Shelley ledger

Fields

∷ ShelleyGenesis StandardShelley

→ Nonce

The initial nonce, typically derived from the hash of Genesis config JSON file.

WARNING: chains using different values of this parameter will be mutually incompatible.

→ ProtVer

→ MaxMajorProtVer

→ Maybe (TPraosLeaderCredentials StandardShelley)

→ Protocol m (ShelleyBlockHFC StandardShelley) ProtocolShelley

ProtocolCardano
Run the protocols of the Cardano block

Fields

∷ Config

→ Maybe PBftSignatureThreshold

→ ProtocolVersion

→ SoftwareVersion

→ Maybe ByronLeaderCredentials

→ ShelleyGenesis StandardShelley

→ Nonce

The initial nonce for the Shelley era, typically derived from the hash of Shelley Genesis config JSON file.

WARNING: chains using different values of this parameter will be mutually incompatible.

→ ProtVer

→ MaxMajorProtVer

→ Maybe (TPraosLeaderCredentials StandardShelley)

→ Maybe EpochNo

maybe lower bound on first Shelley epoch

Setting this to Just when a true lower bound is known may particularly improve performance of bulk syncing. For example, Just 180 would be sound for the Cardano mainnet, since the Byron era's immutable prefix now includes that era. We can update it over time, and set it to the precise value once the transition has actually taken place.

The Nothing case is useful for test and possible alternative nets.

→ TriggerHardFork

→ Protocol m (CardanoBlock StandardCrypto) ProtocolCardano

verifyProtocol ∷ Protocol m blk p → p :~: BlockProtocol blk

Data required to run a protocol

protocolInfo ∷ ∀ m blk p. IOLike m ⇒ Protocol m blk p → ProtocolInfo m blk

Data required to run the selected protocol

Evidence that we can run all the supported protocols

runProtocol ∷ Protocol m blk p → Dict (RunNode blk)

module Ouroboros.Consensus.Shelley.Node

module Ouroboros.Consensus.Byron.Node

module Ouroboros.Consensus.Protocol.PBFT

module Ouroboros.Consensus.Protocol.Abstract

Client support for nodes running a protocol

data ProtocolClient blk p where

Node client support for each consensus protocol.

This is like Protocol but for clients of the node, so with less onerous requirements than to run a node.

Constructors

ProtocolClientByron ∷ EpochSlots → ProtocolClient ByronBlockHFC ProtocolByron

ProtocolClientShelley ∷ ProtocolClient (ShelleyBlockHFC StandardShelley) ProtocolShelley

ProtocolClientCardano ∷ EpochSlots → ProtocolClient (CardanoBlock StandardCrypto) ProtocolCardano

protocolClientInfo ∷ ProtocolClient blk p → ProtocolClientInfo blk

Data required by clients of a node running the specified protocol.

runProtocolClient ∷ ProtocolClient blk p → Dict (RunNode blk)

Sanity check that we have the right class instances available

verifyProtocolClient ∷ ProtocolClient blk p → p :~: BlockProtocol blk

Sanity check that we have the right type combinations