Ouroboros.Consensus.Cardano

The block type of the Cardano block chain

type CardanoBlock c = DegenFork (ShelleyBlock c)

Temporary: use the HardFork combinator with a single block (DegenFork), ShelleyBlock, as it is easier to generate transactions for Shelley than for Byron. This makes sure we exercise all parts of the HardFork combinator in the ThreadNet tests.

We parameterise over crypto (TPraosStandardCrypto for the real block, TPraosMockCrypto for the tests), because we can only generate transactions in the tests when using mock crypto.

Supported protocols

type ProtocolMockBFT = Bft BftMockCrypto

type ProtocolMockPraos = Praos PraosMockCrypto

type ProtocolLeaderSchedule = WithLeaderSchedule (Praos PraosCryptoUnused)

type ProtocolMockPBFT = PBft PBftMockCrypto

type ProtocolRealPBFT = PBft PBftByronCrypto

type ProtocolRealTPraos = TPraos TPraosStandardCrypto

type ProtocolCardano = HardForkProtocol '[ShelleyBlock TPraosStandardCrypto]

Abstract over the various protocols

data Protocol blk p where

Consensus protocol to use

Constructors

ProtocolMockBFT ∷ NumCoreNodes → CoreNodeId → SecurityParam → EraParams → Protocol MockBftBlock ProtocolMockBFT
Run BFT against the mock ledger

ProtocolMockPraos ∷ NumCoreNodes → CoreNodeId → PraosParams → EraParams → Protocol MockPraosBlock ProtocolMockPraos
Run Praos against the mock ledger

ProtocolLeaderSchedule ∷ NumCoreNodes → CoreNodeId → PraosParams → EraParams → LeaderSchedule → Protocol MockPraosRuleBlock ProtocolLeaderSchedule
Run Praos against the mock ledger but with an explicit leader schedule

ProtocolMockPBFT ∷ PBftParams → EraParams → CoreNodeId → Protocol MockPBftBlock ProtocolMockPBFT
Run PBFT against the mock ledger

ProtocolRealPBFT ∷ Config → Maybe PBftSignatureThreshold → ProtocolVersion → SoftwareVersion → Maybe PBftLeaderCredentials → Protocol ByronBlock ProtocolRealPBFT
Run PBFT against the real Byron ledger

ProtocolRealTPraos ∷ ShelleyGenesis TPraosStandardCrypto → ProtVer → Maybe (TPraosLeaderCredentials TPraosStandardCrypto) → Protocol (ShelleyBlock TPraosStandardCrypto) ProtocolRealTPraos
Run TPraos against the real Shelley ledger

ProtocolCardano ∷ ShelleyGenesis TPraosStandardCrypto → ProtVer → Maybe (TPraosLeaderCredentials TPraosStandardCrypto) → Protocol (CardanoBlock TPraosStandardCrypto) ProtocolCardano
Run the protocols of the Cardano block

verifyProtocol ∷ Protocol blk p → p :~: BlockProtocol blk

Data required to run a protocol

protocolInfo ∷ Protocol blk p → ProtocolInfo blk

Data required to run the selected protocol

Evidence that we can run all the supported protocols

runProtocol ∷ Protocol blk p → Dict (RunNode blk)

module Ouroboros.Consensus.Shelley.Node

module Ouroboros.Consensus.Byron.Node

module Ouroboros.Consensus.Mock.Protocol.Praos

module Ouroboros.Consensus.Mock.Node.PraosRule

module Ouroboros.Consensus.Mock.Node.Praos

module Ouroboros.Consensus.Mock.Node.PBFT

module Ouroboros.Consensus.Mock.Node.BFT

module Ouroboros.Consensus.Protocol.PBFT

module Ouroboros.Consensus.Protocol.LeaderSchedule

module Ouroboros.Consensus.Protocol.BFT

module Ouroboros.Consensus.Protocol.Abstract

Client support for nodes running a protocol

data ProtocolClient blk p where

Node client support for each consensus protocol.

This is like Protocol but for clients of the node, so with less onerous requirements than to run a node.

Constructors

ProtocolClientRealPBFT ∷ EpochSlots → ProtocolClient ByronBlock ProtocolRealPBFT

ProtocolClientRealTPraos ∷ ProtocolClient (ShelleyBlock TPraosStandardCrypto) ProtocolRealTPraos

ProtocolClientCardano ∷ ProtocolClient (CardanoBlock TPraosStandardCrypto) ProtocolCardano

protocolClientInfo ∷ ProtocolClient blk p → ProtocolClientInfo blk

Data required by clients of a node running the specified protocol.

runProtocolClient ∷ ProtocolClient blk p → Dict (RunNode blk)

Sanity check that we have the right class instances available

verifyProtocolClient ∷ ProtocolClient blk p → p :~: BlockProtocol blk

Sanity check that we have the right type combinations