Ouroboros.Consensus.Cardano.Block

Eras

type ShelleyEra c = Shelley c

The Shelley ledger and block type itself is parameterised by an era parameter, which is in its turn parameterised by the crypto used.

type CardanoEras c = '[ByronBlock, ShelleyBlock (ShelleyEra c)]

First we had the Byron era and then the Shelley era.

We parameterise over the crypto used in the Shelley era: c.

TODO: parameterise ByronBlock over crypto too

Block

type CardanoBlock c = HardForkBlock (CardanoEras c)

The Cardano block.

Thanks to the pattern synonyms, you can treat this as a sum type with constructors BlockByron and BlockShelley.

toEither :: CardanoBlock c -> Either ByronBlock (ShelleyBlock (ShelleyEra c))
toEither (BlockByron   b) = Left b
toEither (BlockShelley s) = Right s

data HardForkBlock (xs ∷ [Type]) where

Bundled Patterns

pattern BlockByron ∷ ByronBlock → CardanoBlock c

pattern BlockShelley ∷ ShelleyBlock (ShelleyEra c) → CardanoBlock c

Headers

type CardanoHeader c = Header (CardanoBlock c)

The Cardano header.

Thanks to the pattern synonyms, you can treat this as a sum type with constructors HeaderByron and HeaderShelley.

data family Header blk ∷ Type

Generalised transactions

type CardanoGenTx c = GenTx (CardanoBlock c)

The Cardano transaction.

Thanks to the pattern synonyms, you can treat this as a sum type with constructors GenTxByron and GenTxShelley.

data family GenTx blk ∷ Type

Generalized transaction

The mempool (and, accordingly, blocks) consist of "generalized transactions"; this could be "proper" transactions (transferring funds) but also other kinds of things such as update proposals, delegations, etc.

type CardanoGenTxId c = GenTxId (CardanoBlock c)

The ID of a Cardano transaction.

Thanks to the pattern synonyms, you can treat this as a sum type with constructors GenTxIdByron and GenTxIdShelley.

data family TxId tx ∷ Type

A generalized transaction, GenTx, identifier.

type CardanoApplyTxErr c = HardForkApplyTxErr (CardanoEras c)

An error resulting from applying a CardanoGenTx to the ledger.

Thanks to the pattern synonyms, you can treat this as a sum type with constructors ApplyTxByronErr, ApplyTxErrShelley, and ApplyTxErrWrongEra.

toText :: CardanoApplyTxErr c -> Text
toText (ApplyTxErrByron b) = byronApplyTxErrToText b
toText (ApplyTxErrShelley s) = shelleyApplyTxErrToText s
toText (ApplyTxErrWrongEra eraMismatch) =
  "Transaction from the " <> otherEraName eraMismatch <>
  " era applied to a ledger from the " <>
  ledgerEraName eraMismatch <> " era"

data HardForkApplyTxErr (xs ∷ [Type]) where

Bundled Patterns

pattern ApplyTxErrByron ∷ ApplyTxErr ByronBlock → CardanoApplyTxErr c

pattern ApplyTxErrShelley ∷ ApplyTxErr (ShelleyBlock (ShelleyEra c)) → CardanoApplyTxErr c

pattern ApplyTxErrWrongEra ∷ EraMismatch → CardanoApplyTxErr c

LedgerError

type CardanoLedgerError c = HardForkLedgerError (CardanoEras c)

An error resulting from applying a CardanoBlock to the ledger.

Thanks to the pattern synonyms, you can treat this as a sum type with constructors LedgerErrorByron, LedgerErrorShelley, and LedgerErrorWrongEra.

toText :: CardanoLedgerError c -> Text
toText (LedgerErrorByron b) = byronLedgerErrorToText b
toText (LedgerErrorShelley s) = shelleyLedgerErrorToText s
toText (LedgerErrorWrongEra eraMismatch) =
  "Block from the " <> otherEraName eraMismatch <>
  " era applied to a ledger from the " <>
  ledgerEraName eraMismatch <> " era"

data HardForkLedgerError (xs ∷ [Type]) where

Bundled Patterns

pattern LedgerErrorByron ∷ LedgerError ByronBlock → CardanoLedgerError c

pattern LedgerErrorShelley ∷ LedgerError (ShelleyBlock (ShelleyEra c)) → CardanoLedgerError c

pattern LedgerErrorWrongEra ∷ EraMismatch → CardanoLedgerError c

OtherEnvelopeError

type CardanoOtherHeaderEnvelopeError c = HardForkEnvelopeErr (CardanoEras c)

An error resulting from validating a CardanoHeader.

Thanks to the pattern synonyms, you can treat this as a sum type with constructors OtherHeaderEnvelopeErrorByron, OtherHeaderEnvelopeErrorShelley, and OtherHeaderEnvelopeErrorWrongEra.

data HardForkEnvelopeErr (xs ∷ [Type]) where

Bundled Patterns

pattern OtherHeaderEnvelopeErrorByron ∷ OtherHeaderEnvelopeError ByronBlock → CardanoOtherHeaderEnvelopeError c

pattern OtherHeaderEnvelopeErrorShelley ∷ OtherHeaderEnvelopeError (ShelleyBlock (ShelleyEra c)) → CardanoOtherHeaderEnvelopeError c

pattern OtherHeaderEnvelopeErrorWrongEra ∷ EraMismatch → CardanoOtherHeaderEnvelopeError c

TipInfo

type CardanoTipInfo c = OneEraTipInfo (CardanoEras c)

The TipInfo of the Cardano chain.

Thanks to the pattern synonyms, you can treat this as a sum type with constructors TipInfoByron and TipInfoShelley.

data OneEraTipInfo (xs ∷ [Type]) where

Bundled Patterns

pattern TipInfoByron ∷ TipInfo ByronBlock → CardanoTipInfo c

pattern TipInfoShelley ∷ TipInfo (ShelleyBlock (ShelleyEra c)) → CardanoTipInfo c

Query

type CardanoQuery c = Query (CardanoBlock c)

The Query of Cardano chain.

Thanks to the pattern synonyms, you can treat this as a sum type with constructors QueryIfCurrentByron, QueryIfCurrentShelley, QueryAnytimeByron, QueryAnytimeShelley, and Query.

data family Query blk a ∷ Type

Different queries supported by the ledger, indexed by the result type.

type CardanoQueryResult c = HardForkQueryResult (CardanoEras c)

The result of a CardanoQuery

Thanks to the pattern synonyms, you can treat this as a sum type with constructors QueryResultSuccess and QueryResultEraMismatch.

data Either a b where

The Either type represents values with two possibilities: a value of type Either a b is either Left a or Right b.

The Either type is sometimes used to represent a value which is either correct or an error; by convention, the Left constructor is used to hold an error value and the Right constructor is used to hold a correct value (mnemonic: "right" also means "correct").

Examples

Expand

The type Either String Int is the type of values which can be either a String or an Int. The Left constructor can be used only on Strings, and the Right constructor can be used only on Ints:

>>> let s = Left "foo" :: Either String Int
>>> s
Left "foo"
>>> let n = Right 3 :: Either String Int
>>> n
Right 3
>>> :type s
s :: Either String Int
>>> :type n
n :: Either String Int

The fmap from our Functor instance will ignore Left values, but will apply the supplied function to values contained in a Right:

>>> let s = Left "foo" :: Either String Int
>>> let n = Right 3 :: Either String Int
>>> fmap (*2) s
Left "foo"
>>> fmap (*2) n
Right 6

The Monad instance for Either allows us to chain together multiple actions which may fail, and fail overall if any of the individual steps failed. First we'll write a function that can either parse an Int from a Char, or fail.

>>> import Data.Char ( digitToInt, isDigit )
>>> :{
    let parseEither :: Char -> Either String Int
        parseEither c
          | isDigit c = Right (digitToInt c)
          | otherwise = Left "parse error"
>>> :}

The following should work, since both '1' and '2' can be parsed as Ints.

>>> :{
    let parseMultiple :: Either String Int
        parseMultiple = do
          x <- parseEither '1'
          y <- parseEither '2'
          return (x + y)
>>> :}

>>> parseMultiple
Right 3

But the following should fail overall, since the first operation where we attempt to parse 'm' as an Int will fail:

>>> :{
    let parseMultiple :: Either String Int
        parseMultiple = do
          x <- parseEither 'm'
          y <- parseEither '2'
          return (x + y)
>>> :}

>>> parseMultiple
Left "parse error"

Bundled Patterns

pattern QueryResultSuccess ∷ result → CardanoQueryResult c result

pattern QueryResultEraMismatch ∷ EraMismatch → CardanoQueryResult c result
A query from a different era than the ledger's era was sent.

CodecConfig

type CardanoCodecConfig c = CodecConfig (CardanoBlock c)

The CodecConfig for CardanoBlock.

Thanks to the pattern synonyms, you can treat this as the product of the Byron and Shelley CodecConfig.

data family CodecConfig blk ∷ Type

Static configuration required for serialisation and deserialisation of types pertaining to this type of block.

Data family instead of type family to get better type inference.

BlockConfig

type CardanoBlockConfig c = BlockConfig (CardanoBlock c)

The BlockConfig for CardanoBlock.

Thanks to the pattern synonyms, you can treat this as the product of the Byron and Shelley BlockConfig.

data family BlockConfig blk ∷ Type

Static configuration required to work with this type of blocks

ConsensusConfig

type CardanoConsensusConfig c = ConsensusConfig (HardForkProtocol (CardanoEras c))

The ConsensusConfig for CardanoBlock.

Thanks to the pattern synonyms, you can treat this as the product of the Byron and Shelley PartialConsensusConfig.

NOTE: not ConsensusConfig, but PartialConsensusConfig.

data family ConsensusConfig p ∷ Type

Static configuration required to run the consensus protocol

Every method in the ConsensusProtocol class takes the consensus configuration as a parameter, so having this as a data family rather than a type family resolves most ambiguity.

Defined out of the class so that protocols can define this type without having to define the entire protocol at the same time (or indeed in the same module).

LedgerConfig

type CardanoLedgerConfig c = HardForkLedgerConfig (CardanoEras c)

The LedgerConfig for CardanoBlock.

Thanks to the pattern synonyms, you can treat this as the product of the Byron and Shelley PartialLedgerConfig.

NOTE: not LedgerConfig, but PartialLedgerConfig.

data HardForkLedgerConfig (xs ∷ [Type]) where

Bundled Patterns

pattern CardanoLedgerConfig ∷ PartialLedgerConfig ByronBlock → PartialLedgerConfig (ShelleyBlock (ShelleyEra c)) → CardanoLedgerConfig c

LedgerState

type CardanoLedgerState c = LedgerState (CardanoBlock c)

The LedgerState for CardanoBlock.

Thanks to the pattern synonyms, you can inspect the current era's LedgerState using the LedgerStateByron and LedgerStateShelley constructors.

NOTE: the CardanoLedgerState contains more than just the current era's LedgerState. We don't give access to those internal details through the pattern synonyms. This is also the reason the pattern synonyms are not bidirectional.

data family LedgerState blk ∷ Type

Ledger state associated with a block

EraMismatch

data EraMismatch

Extra info for errors caused by applying a block, header, transaction, or query from one era to a ledger from a different era.

Constructors

EraMismatch

Fields

ledgerEraName ∷ !Text

Name of the era of the ledger (Byron or Shelley).

otherEraName ∷ !Text

Era of the block, header, transaction, or query.