Ouroboros.Consensus.Cardano.Block

Documentation

type CardanoBlock c = DegenFork (ShelleyBlock c)

Temporary: use the HardFork combinator with a single block (DegenFork), ShelleyBlock, as it is easier to generate transactions for Shelley than for Byron. This makes sure we exercise all parts of the HardFork combinator in the ThreadNet tests.

We parameterise over crypto (TPraosStandardCrypto for the real block, TPraosMockCrypto for the tests), because we can only generate transactions in the tests when using mock crypto.

Orphan instances

TPraosCrypto c ⇒ SingleEraBlock (ShelleyBlock c)

Instance details

Methods

singleEraParams ∷ proxy (ShelleyBlock c) → PartialLedgerConfig (ShelleyBlock c) → EraParams

singleEraTransition ∷ PartialLedgerConfig (ShelleyBlock c) → LedgerState (ShelleyBlock c) → Maybe EpochNo

singleEraInfo ∷ proxy (ShelleyBlock c) → SingleEraInfo (ShelleyBlock c)

TPraosCrypto c ⇒ Condense (CardanoBlock c)

Instance details

Methods

condense ∷ CardanoBlock c → String

TPraosCrypto c ⇒ HasPartialConsensusConfig (TPraos c)

Instance details

Associated Types

type PartialConsensusConfig (TPraos c) ∷ Type

Methods

completeConsensusConfig ∷ proxy (TPraos c) → EpochInfo Identity → PartialConsensusConfig (TPraos c) → ConsensusConfig (TPraos c)

partialChainSelConfig ∷ proxy (TPraos c) → PartialConsensusConfig (TPraos c) → ChainSelConfig (TPraos c)

partialCheckIfCanBeLeader ∷ proxy (TPraos c) → PartialConsensusConfig (TPraos c) → Bool

TPraosCrypto c ⇒ HasPartialLedgerConfig (ShelleyBlock c)

Instance details

Associated Types

type PartialLedgerConfig (ShelleyBlock c) ∷ Type

Methods

completeLedgerConfig ∷ proxy (ShelleyBlock c) → EpochInfo Identity → PartialLedgerConfig (ShelleyBlock c) → LedgerConfig (ShelleyBlock c)