Ouroboros.Consensus.Cardano.ByronHFC

Documentation

type ByronBlockHFC = HardForkBlock '[ByronBlock]

Byron as the single era in the hard fork combinator

Orphan instances

NoHardForks ByronBlock

Instance details

Methods

getEraParams ∷ TopLevelConfig ByronBlock → EraParams

toPartialConsensusConfig ∷ proxy ByronBlock → ConsensusConfig (BlockProtocol ByronBlock) → PartialConsensusConfig (BlockProtocol ByronBlock)

toPartialLedgerConfig ∷ proxy ByronBlock → LedgerConfig ByronBlock → PartialLedgerConfig ByronBlock

SupportedNetworkProtocolVersion ByronBlockHFC
Forward to the ByronBlock instance. Only supports HardForkNodeToNodeDisabled, which is compatible with nodes running with ByronBlock.

Instance details

Methods

supportedNodeToNodeVersions ∷ Proxy ByronBlockHFC → Map NodeToNodeVersion (BlockNodeToNodeVersion ByronBlockHFC)

supportedNodeToClientVersions ∷ Proxy ByronBlockHFC → Map NodeToClientVersion (BlockNodeToClientVersion ByronBlockHFC)

latestReleasedNodeVersion ∷ Proxy ByronBlockHFC → (Maybe NodeToNodeVersion, Maybe NodeToClientVersion)

SerialiseHFC '[ByronBlock]
Forward to the ByronBlock instance, this means we don't add an era wrapper around blocks on disk. This makes sure we're compatible with the existing Byron blocks.

Instance details

Methods

encodeDiskHfcBlock ∷ CodecConfig (HardForkBlock '[ByronBlock]) → HardForkBlock '[ByronBlock] → Encoding

decodeDiskHfcBlock ∷ CodecConfig (HardForkBlock '[ByronBlock]) → ∀ s. Decoder s (ByteString → HardForkBlock '[ByronBlock])

reconstructHfcPrefixLen ∷ proxy (Header (HardForkBlock '[ByronBlock])) → PrefixLen

reconstructHfcNestedCtxt ∷ proxy (Header (HardForkBlock '[ByronBlock])) → ShortByteString → SizeInBytes → SomeSecond (NestedCtxt Header) (HardForkBlock '[ByronBlock])

getHfcBinaryBlockInfo ∷ HardForkBlock '[ByronBlock] → BinaryBlockInfo

estimateHfcBlockSize ∷ Header (HardForkBlock '[ByronBlock]) → SizeInBytes