Ouroboros.Consensus.Cardano.CanHardFork

Documentation

data HardCodedTransition

Constructors

NoHardCodedTransition

HardCodedTransitionAt !EpochNo

data ByronPartialLedgerConfig

When Byron is part of the hard-fork combinator, we use the partial ledger config. Standalone Byron uses the regular ledger config. This means that the partial ledger config is the perfect place to store a hard-coded, yet configurable, transition point for testing purposes, as we don't have to modify the ledger config for standalone Byron.

Constructors

ByronPartialLedgerConfig

Fields

byronLedgerConfig ∷ !(LedgerConfig ByronBlock)

transitionEpoch ∷ !HardCodedTransition

data ShelleyPartialLedgerConfig sc

Constructors

ShelleyPartialLedgerConfig

Fields

shelleyPartialLedgerGenesis ∷ !(ShelleyGenesis sc)

shelleyPartialMaxMajorPV ∷ !Natural

Orphan instances

SingleEraBlock ByronBlock

Instance details

Methods

singleEraTransition ∷ PartialLedgerConfig ByronBlock → LedgerState ByronBlock → Maybe EpochNo

singleEraInfo ∷ proxy ByronBlock → SingleEraInfo ByronBlock

HasPartialLedgerConfig ByronBlock

Instance details

Associated Types

type PartialLedgerConfig ByronBlock ∷ Type

Methods

completeLedgerConfig ∷ proxy ByronBlock → EpochInfo Identity → PartialLedgerConfig ByronBlock → LedgerConfig ByronBlock

TPraosCrypto c ⇒ CanHardFork (CardanoEras c)

Instance details

Methods

hardForkEraTranslation ∷ EraTranslation (CardanoEras c)

TPraosCrypto sc ⇒ SingleEraBlock (ShelleyBlock sc)

Instance details

Methods

singleEraTransition ∷ PartialLedgerConfig (ShelleyBlock sc) → LedgerState (ShelleyBlock sc) → Maybe EpochNo

singleEraInfo ∷ proxy (ShelleyBlock sc) → SingleEraInfo (ShelleyBlock sc)

PBftCrypto bc ⇒ HasPartialConsensusConfig (PBft bc)

Instance details

Associated Types

type PartialConsensusConfig (PBft bc) ∷ Type

Methods

completeConsensusConfig ∷ proxy (PBft bc) → EpochInfo Identity → PartialConsensusConfig (PBft bc) → ConsensusConfig (PBft bc)

partialChainSelConfig ∷ proxy (PBft bc) → PartialConsensusConfig (PBft bc) → ChainSelConfig (PBft bc)

TPraosCrypto sc ⇒ HasPartialConsensusConfig (TPraos sc)

Instance details

Associated Types

type PartialConsensusConfig (TPraos sc) ∷ Type

Methods

completeConsensusConfig ∷ proxy (TPraos sc) → EpochInfo Identity → PartialConsensusConfig (TPraos sc) → ConsensusConfig (TPraos sc)

partialChainSelConfig ∷ proxy (TPraos sc) → PartialConsensusConfig (TPraos sc) → ChainSelConfig (TPraos sc)

TPraosCrypto sc ⇒ HasPartialLedgerConfig (ShelleyBlock sc)

Instance details

Associated Types

type PartialLedgerConfig (ShelleyBlock sc) ∷ Type

Methods

completeLedgerConfig ∷ proxy (ShelleyBlock sc) → EpochInfo Identity → PartialLedgerConfig (ShelleyBlock sc) → LedgerConfig (ShelleyBlock sc)