Ouroboros.Consensus.Cardano.CanHardFork

Documentation

data TriggerHardFork

The trigger condition that will cause the hard fork transition.

Constructors

TriggerHardForkAtVersion !Word16
Trigger the transition when the on-chain protocol major version (from the ledger state) reaches this number.

TriggerHardForkAtEpoch !EpochNo
For testing only, trigger the transition at a specific hard-coded epoch, irrespective of the ledger state.

TriggerHardForkNever
Never trigger a hard fork

data ByronPartialLedgerConfig

When Byron is part of the hard-fork combinator, we use the partial ledger config. Standalone Byron uses the regular ledger config. This means that the partial ledger config is the perfect place to store the trigger condition for the hard fork to Shelley, as we don't have to modify the ledger config for standalone Byron.

Constructors

ByronPartialLedgerConfig

Fields

byronLedgerConfig ∷ !(LedgerConfig ByronBlock)

byronTriggerHardFork ∷ !TriggerHardFork

data ShelleyPartialLedgerConfig era

Constructors

ShelleyPartialLedgerConfig

Fields

shelleyLedgerConfig ∷ !(ShelleyLedgerConfig era)

We cache the non-partial ledger config containing a dummy EpochInfo that needs to be replaced with the correct one.

We do this to avoid recomputing the ledger config each time completeLedgerConfig is called, as mkShelleyLedgerConfig does some rather expensive computations that shouldn't be repeated too often (e.g., sgActiveSlotCoeff).

shelleyTriggerHardFork ∷ !TriggerHardFork

type CardanoHardForkConstraints c = (TPraosCrypto c, ShelleyBasedEra (ShelleyEra c), ShelleyBasedEra (AllegraEra c), ShelleyBasedEra (MaryEra c), HASH c ~ Blake2b_256, ADDRHASH c ~ Blake2b_224, DSIGN c ~ Ed25519DSIGN)

Orphan instances

SingleEraBlock ByronBlock

Instance details

Methods

singleEraTransition ∷ PartialLedgerConfig ByronBlock → EraParams → Bound → LedgerState ByronBlock → Maybe EpochNo

singleEraInfo ∷ proxy ByronBlock → SingleEraInfo ByronBlock

HasPartialLedgerConfig ByronBlock

Instance details

Associated Types

type PartialLedgerConfig ByronBlock

Methods

completeLedgerConfig ∷ proxy ByronBlock → EpochInfo Identity → PartialLedgerConfig ByronBlock → LedgerConfig ByronBlock

(ShelleyBasedEra era, ShelleyBasedEra (PreviousEra era), EraCrypto (PreviousEra era) ~ EraCrypto era) ⇒ TranslateEra era ShelleyTip

Instance details

Associated Types

type TranslationError era ShelleyTip

Methods

translateEra ∷ TranslationContext era → ShelleyTip (PreviousEra era) → Except (TranslationError era ShelleyTip) (ShelleyTip era)

(ShelleyBasedEra era, TranslateEra era Tx) ⇒ TranslateEra era (GenTx :.: ShelleyBlock)

Instance details

Associated Types

type TranslationError era (GenTx :.: ShelleyBlock)

Methods

translateEra ∷ TranslationContext era → (GenTx :.: ShelleyBlock) (PreviousEra era) → Except (TranslationError era (GenTx :.: ShelleyBlock)) ((GenTx :.: ShelleyBlock) era)

(ShelleyBasedEra era, TranslateEra era ShelleyTip, TranslateEra era NewEpochState, TranslationError era NewEpochState ~ Void) ⇒ TranslateEra era (LedgerState :.: ShelleyBlock)

Instance details

Associated Types

type TranslationError era (LedgerState :.: ShelleyBlock)

Methods

translateEra ∷ TranslationContext era → (LedgerState :.: ShelleyBlock) (PreviousEra era) → Except (TranslationError era (LedgerState :.: ShelleyBlock)) ((LedgerState :.: ShelleyBlock) era)

CardanoHardForkConstraints c ⇒ CanHardFork (CardanoEras c)

Instance details

Methods

hardForkEraTranslation ∷ EraTranslation (CardanoEras c)

hardForkChainSel ∷ Tails AcrossEraSelection (CardanoEras c)

hardForkInjectTxs ∷ InPairs (RequiringBoth WrapLedgerConfig InjectTx) (CardanoEras c)

ShelleyBasedEra era ⇒ SingleEraBlock (ShelleyBlock era)

Instance details

Methods

singleEraTransition ∷ PartialLedgerConfig (ShelleyBlock era) → EraParams → Bound → LedgerState (ShelleyBlock era) → Maybe EpochNo

singleEraInfo ∷ proxy (ShelleyBlock era) → SingleEraInfo (ShelleyBlock era)

PBftCrypto bc ⇒ HasPartialConsensusConfig (PBft bc)

Instance details

Associated Types

type PartialConsensusConfig (PBft bc)

Methods

completeConsensusConfig ∷ proxy (PBft bc) → EpochInfo Identity → PartialConsensusConfig (PBft bc) → ConsensusConfig (PBft bc)

partialChainSelConfig ∷ proxy (PBft bc) → PartialConsensusConfig (PBft bc) → ChainSelConfig (PBft bc)

TPraosCrypto c ⇒ HasPartialConsensusConfig (TPraos c)

Instance details

Associated Types

type PartialConsensusConfig (TPraos c)

Methods

completeConsensusConfig ∷ proxy (TPraos c) → EpochInfo Identity → PartialConsensusConfig (TPraos c) → ConsensusConfig (TPraos c)

partialChainSelConfig ∷ proxy (TPraos c) → PartialConsensusConfig (TPraos c) → ChainSelConfig (TPraos c)

ShelleyBasedEra era ⇒ HasPartialLedgerConfig (ShelleyBlock era)

Instance details

Associated Types

type PartialLedgerConfig (ShelleyBlock era)

Methods

completeLedgerConfig ∷ proxy (ShelleyBlock era) → EpochInfo Identity → PartialLedgerConfig (ShelleyBlock era) → LedgerConfig (ShelleyBlock era)