Ouroboros.Consensus.Cardano.CanHardFork

Documentation

data TriggerHardFork

The trigger condition that will cause the hard fork transition.

Constructors

TriggerHardForkAtVersion !Word16
Trigger the transition when the on-chain protocol major version (from the ledger state) reaches this number.

TriggerHardForkAtEpoch !EpochNo
For testing only, trigger the transition at a specific hard-coded epoch, irrespective of the ledger state.

TriggerHardForkNever
Never trigger a hard fork

data ByronPartialLedgerConfig

When Byron is part of the hard-fork combinator, we use the partial ledger config. Standalone Byron uses the regular ledger config. This means that the partial ledger config is the perfect place to store the trigger condition for the hard fork to Shelley, as we don't have to modify the ledger config for standalone Byron.

Constructors

ByronPartialLedgerConfig

Fields

byronLedgerConfig ∷ !(LedgerConfig ByronBlock)

triggerHardFork ∷ !TriggerHardFork

newtype ShelleyPartialLedgerConfig era

Constructors

ShelleyPartialLedgerConfig

Fields

getShelleyPartialLedgerConfig ∷ ShelleyLedgerConfig era

We cache the non-partial ledger config containing a dummy EpochInfo that needs to be replaced with the correct one.

We do this to avoid recomputing the ledger config each time completeLedgerConfig is called, as mkShelleyLedgerConfig does some rather expensive computations that shouldn't be repeated too often (e.g., sgActiveSlotCoeff).

type CardanoHardForkConstraints c = (TPraosCrypto (ShelleyEra c), HASH c ~ Blake2b_256, ADDRHASH c ~ Blake2b_224, DSIGN c ~ Ed25519DSIGN)

Orphan instances

SingleEraBlock ByronBlock

Instance details

Methods

singleEraTransition ∷ PartialLedgerConfig ByronBlock → EraParams → Bound → LedgerState ByronBlock → Maybe EpochNo

singleEraInfo ∷ proxy ByronBlock → SingleEraInfo ByronBlock

HasPartialLedgerConfig ByronBlock

Instance details

Associated Types

type PartialLedgerConfig ByronBlock ∷ Type

Methods

completeLedgerConfig ∷ proxy ByronBlock → EpochInfo Identity → PartialLedgerConfig ByronBlock → LedgerConfig ByronBlock

CardanoHardForkConstraints c ⇒ CanHardFork (CardanoEras c)

Instance details

Methods

hardForkEraTranslation ∷ EraTranslation (CardanoEras c)

hardForkChainSel ∷ Tails AcrossEraSelection (CardanoEras c)

hardForkInjectTxs ∷ InPairs (RequiringBoth WrapLedgerConfig InjectTx) (CardanoEras c)

TPraosCrypto era ⇒ SingleEraBlock (ShelleyBlock era)

Instance details

Methods

singleEraTransition ∷ PartialLedgerConfig (ShelleyBlock era) → EraParams → Bound → LedgerState (ShelleyBlock era) → Maybe EpochNo

singleEraInfo ∷ proxy (ShelleyBlock era) → SingleEraInfo (ShelleyBlock era)

PBftCrypto bc ⇒ HasPartialConsensusConfig (PBft bc)

Instance details

Associated Types

type PartialConsensusConfig (PBft bc) ∷ Type

Methods

completeConsensusConfig ∷ proxy (PBft bc) → EpochInfo Identity → PartialConsensusConfig (PBft bc) → ConsensusConfig (PBft bc)

partialChainSelConfig ∷ proxy (PBft bc) → PartialConsensusConfig (PBft bc) → ChainSelConfig (PBft bc)

TPraosCrypto era ⇒ HasPartialConsensusConfig (TPraos era)

Instance details

Associated Types

type PartialConsensusConfig (TPraos era) ∷ Type

Methods

completeConsensusConfig ∷ proxy (TPraos era) → EpochInfo Identity → PartialConsensusConfig (TPraos era) → ConsensusConfig (TPraos era)

partialChainSelConfig ∷ proxy (TPraos era) → PartialConsensusConfig (TPraos era) → ChainSelConfig (TPraos era)

TPraosCrypto era ⇒ HasPartialLedgerConfig (ShelleyBlock era)

Instance details

Associated Types

type PartialLedgerConfig (ShelleyBlock era) ∷ Type

Methods

completeLedgerConfig ∷ proxy (ShelleyBlock era) → EpochInfo Identity → PartialLedgerConfig (ShelleyBlock era) → LedgerConfig (ShelleyBlock era)