Ouroboros.Consensus.Cardano.CanHardFork

Documentation

data ByronPartialLedgerConfig

When Byron is part of the hard-fork combinator, we use the partial ledger config. Standalone Byron uses the regular ledger config. This means that the partial ledger config is the perfect place to store the trigger condition for the hard fork to Shelley, as we don't have to modify the ledger config for standalone Byron.

Constructors

ByronPartialLedgerConfig

Fields

byronLedgerConfig ∷ !(LedgerConfig ByronBlock)

byronTriggerHardFork ∷ !TriggerHardFork

type CardanoHardForkConstraints c = (PraosCrypto c, ShelleyBasedEra (ShelleyEra c), ShelleyBasedEra (AllegraEra c), ShelleyBasedEra (MaryEra c), ShelleyBasedEra (AlonzoEra c), HASH c ~ Blake2b_256, ADDRHASH c ~ Blake2b_224, DSIGN c ~ Ed25519DSIGN)

data TriggerHardFork

The trigger condition that will cause the hard fork transition.

Constructors

TriggerHardForkAtVersion !Word16
Trigger the transition when the on-chain protocol major version (from the ledger state) reaches this number.

TriggerHardForkAtEpoch !EpochNo
For testing only, trigger the transition at a specific hard-coded epoch, irrespective of the ledger state.

TriggerHardForkNever
Never trigger a hard fork

Re-exports of Shelley code

data ShelleyPartialLedgerConfig era

Constructors

ShelleyPartialLedgerConfig

Fields

shelleyLedgerConfig ∷ !(ShelleyLedgerConfig era)

We cache the non-partial ledger config containing a dummy EpochInfo that needs to be replaced with the correct one.

We do this to avoid recomputing the ledger config each time completeLedgerConfig is called, as mkShelleyLedgerConfig does some rather expensive computations that shouldn't be repeated too often (e.g., sgActiveSlotCoeff).

shelleyTriggerHardFork ∷ !TriggerHardFork

forecastAcrossShelley

Arguments

∷ (EraCrypto eraFrom ~ EraCrypto eraTo, ShelleyBasedEra eraFrom)

⇒ ShelleyLedgerConfig eraFrom

→ ShelleyLedgerConfig eraTo

→ Bound
Transition between the two eras

→ SlotNo
Forecast for this slot

→ LedgerState (ShelleyBlock eraFrom)

→ Except OutsideForecastRange (Ticked (WrapLedgerView (ShelleyBlock eraTo)))

Forecast from a Shelley-based era to the next Shelley-based era.

translateChainDepStateAcrossShelley ∷ EraCrypto eraFrom ~ EraCrypto eraTo ⇒ RequiringBoth WrapConsensusConfig (Translate WrapChainDepState) (ShelleyBlock eraFrom) (ShelleyBlock eraTo)

Orphan instances

SingleEraBlock ByronBlock

Instance details

Methods

singleEraTransition ∷ PartialLedgerConfig ByronBlock → EraParams → Bound → LedgerState ByronBlock → Maybe EpochNo

singleEraInfo ∷ proxy ByronBlock → SingleEraInfo ByronBlock

HasPartialLedgerConfig ByronBlock

Instance details

Associated Types

type PartialLedgerConfig ByronBlock

Methods

completeLedgerConfig ∷ proxy ByronBlock → EpochInfo (Except PastHorizonException) → PartialLedgerConfig ByronBlock → LedgerConfig ByronBlock

CardanoHardForkConstraints c ⇒ CanHardFork (CardanoEras c)

Instance details

Methods

hardForkEraTranslation ∷ EraTranslation (CardanoEras c)

hardForkChainSel ∷ Tails AcrossEraSelection (CardanoEras c)

hardForkInjectTxs ∷ InPairs (RequiringBoth WrapLedgerConfig (Product2 InjectTx InjectValidatedTx)) (CardanoEras c)

PBftCrypto bc ⇒ HasPartialConsensusConfig (PBft bc)

Instance details

Associated Types

type PartialConsensusConfig (PBft bc)

Methods

completeConsensusConfig ∷ proxy (PBft bc) → EpochInfo (Except PastHorizonException) → PartialConsensusConfig (PBft bc) → ConsensusConfig (PBft bc)