Ouroboros.Consensus.Cardano.Node

Documentation

protocolInfoCardano ∷ ∀ sc m. (MonadRandom m, TPraosCrypto sc) ⇒ Config → Maybe PBftSignatureThreshold → ProtocolVersion → SoftwareVersion → Maybe PBftLeaderCredentials → ShelleyGenesis sc → ProtVer → Natural → Maybe (TPraosLeaderCredentials sc) → HardCodedTransition → ProtocolInfo m (CardanoBlock sc)

protocolClientInfoCardano ∷ ∀ sc. EpochSlots → SecurityParam → ProtocolClientInfo (CardanoBlock sc)

data HardCodedTransition

Constructors

NoHardCodedTransition

HardCodedTransitionAt !EpochNo

TranslateNetworkProtocolVersion

pattern CardanoNodeToNodeVersion1 ∷ BlockNodeToNodeVersion (CardanoBlock sc)

We support only Byron V1 with the hard fork disabled, as no other versions have been released before the hard fork

pattern CardanoNodeToNodeVersion2 ∷ BlockNodeToNodeVersion (CardanoBlock sc)

pattern CardanoNodeToClientVersion1 ∷ BlockNodeToClientVersion (CardanoBlock sc)

We support Byron V1 with hard fork disabled, as it was released before the hard fork

pattern CardanoNodeToClientVersion2 ∷ BlockNodeToClientVersion (CardanoBlock sc)

We support Byron V2 with hard fork disabled, as it was released before the hard fork

pattern CardanoNodeToClientVersion3 ∷ BlockNodeToClientVersion (CardanoBlock sc)

Orphan instances

SerialiseConstraintsHFC ByronBlock

Instance details

TPraosCrypto sc ⇒ SerialiseConstraintsHFC (ShelleyBlock sc)

Instance details

TPraosCrypto sc ⇒ SerialiseHFC (CardanoEras sc)
Important: we need to maintain binary compatibility with Byron blocks, as they are already stored on disk.

We also want to be able to efficiently detect (without having to peek far ahead) whether we're dealing with a Byron or Shelley block, so that we can invoke the right decoder. We plan to have a few more hard forks after Shelley (Goguen, Basho, Voltaire), so we want a future-proof envelope for distinguishing the different block types, i.e., a byte indicating the era.

Byron does not provide such an envelope. However, a Byron block is a CBOR 2-tuple with the first element being a tag (Word: 0 = EBB; 1 = regular block) and the second being the payload. We can easily extend this encoding format with support for Shelley, Goguen, etc.

We encode a CardanoBlock as the same CBOR 2-tuple as a Byron block, but we use the tags after 1 for the hard forks after Byron:

Byron EBB

Byron regular block

Shelley block

Goguen block

etc.

For more details, see: https://github.com/input-output-hk/ouroboros-network/pull/1175#issuecomment-558147194

Instance details

Methods

encodeDiskHfcBlock ∷ CodecConfig (HardForkBlock (CardanoEras sc)) → HardForkBlock (CardanoEras sc) → Encoding

decodeDiskHfcBlock ∷ CodecConfig (HardForkBlock (CardanoEras sc)) → ∀ s. Decoder s (ByteString → HardForkBlock (CardanoEras sc))

reconstructHfcPrefixLen ∷ proxy (Header (HardForkBlock (CardanoEras sc))) → Word8

reconstructHfcNestedCtxt ∷ proxy (Header (HardForkBlock (CardanoEras sc))) → ShortByteString → SizeInBytes → SomeBlock (NestedCtxt Header) (HardForkBlock (CardanoEras sc))

TPraosCrypto sc ⇒ RunNode (CardanoBlock sc)

Instance details

Methods

nodeBlockFetchSize ∷ Header (CardanoBlock sc) → SizeInBytes

nodeImmDbChunkInfo ∷ TopLevelConfig (CardanoBlock sc) → ChunkInfo

nodeCheckIntegrity ∷ TopLevelConfig (CardanoBlock sc) → CardanoBlock sc → Bool

nodeGetBinaryBlockInfo ∷ CardanoBlock sc → BinaryBlockInfo

nodeInitChainDB ∷ IOLike m ⇒ TopLevelConfig (CardanoBlock sc) → InitChainDB m (CardanoBlock sc) → m ()

TPraosCrypto sc ⇒ TranslateNetworkProtocolVersion (CardanoBlock sc)

Instance details

Methods

supportedNodeToNodeVersions ∷ Proxy (CardanoBlock sc) → NonEmpty (BlockNodeToNodeVersion (CardanoBlock sc))

supportedNodeToClientVersions ∷ Proxy (CardanoBlock sc) → NonEmpty (BlockNodeToClientVersion (CardanoBlock sc))

mostRecentSupportedNodeToNode ∷ Proxy (CardanoBlock sc) → BlockNodeToNodeVersion (CardanoBlock sc)

mostRecentSupportedNodeToClient ∷ Proxy (CardanoBlock sc) → BlockNodeToClientVersion (CardanoBlock sc)

nodeToNodeProtocolVersion ∷ Proxy (CardanoBlock sc) → BlockNodeToNodeVersion (CardanoBlock sc) → NodeToNodeVersion

nodeToClientProtocolVersion ∷ Proxy (CardanoBlock sc) → BlockNodeToClientVersion (CardanoBlock sc) → NodeToClientVersion