Ouroboros.Consensus.Cardano.Node

Documentation

protocolInfoCardano

Arguments

∷ (IOLike m, CardanoHardForkConstraints c)

⇒ Config

→ Maybe PBftSignatureThreshold

→ ProtocolVersion

→ SoftwareVersion

→ Maybe ByronLeaderCredentials

→ ShelleyGenesis (ShelleyEra c)

→ Nonce
The initial nonce for the Shelley era, typically derived from the hash of Shelley Genesis config JSON file.

→ ProtVer

→ MaxMajorProtVer

→ Maybe (TPraosLeaderCredentials (ShelleyEra c))

→ Maybe EpochNo
lower bound on first Shelley epoch

→ TriggerHardFork

→ ProtocolInfo m (CardanoBlock c)

protocolClientInfoCardano ∷ ∀ c. EpochSlots → ProtocolClientInfo (CardanoBlock c)

type CardanoHardForkConstraints c = (TPraosCrypto (ShelleyEra c), HASH c ~ Blake2b_256, ADDRHASH c ~ Blake2b_224, DSIGN c ~ Ed25519DSIGN)

newtype MaxMajorProtVer

The maximum major protocol version.

Must be at least the current major protocol version. For Cardano mainnet, the Shelley era has major protocol verison 2.

Constructors

MaxMajorProtVer

Fields

getMaxMajorProtVer ∷ Natural

data TriggerHardFork

The trigger condition that will cause the hard fork transition.

Constructors

TriggerHardForkAtVersion !Word16
Trigger the transition when the on-chain protocol major version (from the ledger state) reaches this number.

TriggerHardForkAtEpoch !EpochNo
For testing only, trigger the transition at a specific hard-coded epoch, irrespective of the ledger state.

TriggerHardForkNever
Never trigger a hard fork

initialLedgerStateCardano ∷ Config → LedgerState (CardanoBlock c)

Create the initial LedgerState based on the given Byron genesis config.

ledgerConfigCardano

Arguments

∷ Config

→ ShelleyGenesis (ShelleyEra c)

→ MaxMajorProtVer

→ TriggerHardFork

→ Maybe EpochNo
lower bound on first Shelley epoch

→ CardanoLedgerConfig c

Create a LedgerConfig for CardanoBlock.

mkPartialLedgerConfigShelley ∷ ShelleyGenesis (ShelleyEra c) → MaxMajorProtVer → PartialLedgerConfig (ShelleyBlock (ShelleyEra c))

SupportedNetworkProtocolVersion

pattern CardanoNodeToNodeVersion1 ∷ BlockNodeToNodeVersion (CardanoBlock c)

We support only Byron V1 with the hard fork disabled, as no other versions have been released before the hard fork

pattern CardanoNodeToNodeVersion2 ∷ BlockNodeToNodeVersion (CardanoBlock c)

pattern CardanoNodeToClientVersion1 ∷ BlockNodeToClientVersion (CardanoBlock c)

We support Byron V1 with hard fork disabled, as it was released before the hard fork

pattern CardanoNodeToClientVersion2 ∷ BlockNodeToClientVersion (CardanoBlock c)

Orphan instances

SerialiseConstraintsHFC ByronBlock

Instance details

TPraosCrypto era ⇒ SerialiseConstraintsHFC (ShelleyBlock era)

Instance details

CardanoHardForkConstraints c ⇒ SerialiseHFC (CardanoEras c)
Important: we need to maintain binary compatibility with Byron blocks, as they are already stored on disk.

We also want to be able to efficiently detect (without having to peek far ahead) whether we're dealing with a Byron or Shelley block, so that we can invoke the right decoder. We plan to have a few more hard forks after Shelley (Goguen, Basho, Voltaire), so we want a future-proof envelope for distinguishing the different block types, i.e., a byte indicating the era.

Byron does not provide such an envelope. However, a Byron block is a CBOR 2-tuple with the first element being a tag (Word: 0 = EBB; 1 = regular block) and the second being the payload. We can easily extend this encoding format with support for Shelley, Goguen, etc.

We encode a CardanoBlock as the same CBOR 2-tuple as a Byron block, but we use the tags after 1 for the hard forks after Byron:

Byron EBB

Byron regular block

Shelley block

Goguen block

etc.

For more details, see: https://github.com/input-output-hk/ouroboros-network/pull/1175#issuecomment-558147194

Instance details

Methods

encodeDiskHfcBlock ∷ CodecConfig (HardForkBlock (CardanoEras c)) → HardForkBlock (CardanoEras c) → Encoding

decodeDiskHfcBlock ∷ CodecConfig (HardForkBlock (CardanoEras c)) → ∀ s. Decoder s (ByteString → HardForkBlock (CardanoEras c))

reconstructHfcPrefixLen ∷ proxy (Header (HardForkBlock (CardanoEras c))) → PrefixLen

reconstructHfcNestedCtxt ∷ proxy (Header (HardForkBlock (CardanoEras c))) → ShortByteString → SizeInBytes → SomeBlock (NestedCtxt Header) (HardForkBlock (CardanoEras c))

getHfcBinaryBlockInfo ∷ HardForkBlock (CardanoEras c) → BinaryBlockInfo

CardanoHardForkConstraints c ⇒ RunNode (CardanoBlock c)

Instance details

Methods

nodeBlockFetchSize ∷ Header (CardanoBlock c) → SizeInBytes

nodeImmutableDbChunkInfo ∷ TopLevelConfig (CardanoBlock c) → ChunkInfo

nodeCheckIntegrity ∷ TopLevelConfig (CardanoBlock c) → CardanoBlock c → Bool

nodeInitChainDB ∷ IOLike m ⇒ TopLevelConfig (CardanoBlock c) → InitChainDB m (CardanoBlock c) → m ()

CardanoHardForkConstraints c ⇒ SupportedNetworkProtocolVersion (CardanoBlock c)

Instance details

Methods

supportedNodeToNodeVersions ∷ Proxy (CardanoBlock c) → Map NodeToNodeVersion (BlockNodeToNodeVersion (CardanoBlock c))

supportedNodeToClientVersions ∷ Proxy (CardanoBlock c) → Map NodeToClientVersion (BlockNodeToClientVersion (CardanoBlock c))